% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_config.R
\name{add_datasource}
\alias{add_datasource}
\title{Add a new datasource to a YAML configuration file}
\usage{
add_datasource(config_path, name, backend)
}
\arguments{
\item{config_path}{The file path to the YAML configuration file}

\item{name}{The name of the new datasource}

\item{backend}{A named list representing the backend configuration for the new datasource}
}
\value{
The updated configuration after adding the new datasource
}
\description{
This function adds a new datasource to a YAML configuration file by appending the
provided datasource information to the existing datasources.
}
