% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_files.R
\name{assert_ext}
\alias{assert_ext}
\title{Test the extension of files}
\usage{
assert_ext(ext, method)
}
\arguments{
\item{ext}{the extension to test}

\item{method}{the S3 method to get methods}
}
\value{
An error if the extension method doesn't exists
}
\description{
Test the extension of files
}
