% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{assert_path}
\alias{assert_path}
\title{Validate the path and access}
\usage{
assert_path(path, access)
}
\arguments{
\item{path}{Path to be validated}

\item{access}{Type of access ("rw" for read/write by default)}
}
\value{
Invisible path
}
\description{
The assert_path function validates the path and access type for file system operations.
}
