% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetweenCluster_CurvDist.R
\name{BetweenCluster_CurvDist}
\alias{BetweenCluster_CurvDist}
\title{Cluster Betweenness}
\usage{
BetweenCluster_CurvDist(ClustCurve, i)
}
\arguments{
\item{ClustCurve}{Data frame with five arguments : time, growth values, ID, cluster membership and feature values for each sample.}

\item{i}{Value included from 0 to K (number of clusters) representing the cluster involved in withinness computation.}
}
\value{
Returns a list with four arguments: (i) CurveDistance, a matrix composed by the hausdorff distance among the samples belonging to the i-th cluster, (ii) Between, a matrix for minimum and maximum Hausdorff distance of i-th cluster curves from the other clusters, (iii) NearCurveID and (iv) FarCurveID are the nearest and the farthest samples from the i-th cluster.
}
\description{
Cluster Betweenness
}
