% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetweenCluster_MeanDist.R
\name{BetweenCluster_MeanDist}
\alias{BetweenCluster_MeanDist}
\title{Centroid betweenness}
\usage{
BetweenCluster_MeanDist(ClustCurve, MeanCurves, i)
}
\arguments{
\item{ClustCurve}{Data frame with five arguments : time, growth values, ID, cluster membership and feature values for each sample.}

\item{MeanCurves}{Matrix with the meancurve values on the columns according to different clusters.}

\item{i}{Value included from 0 to K (number of clusters) representing the cluster involved in withinness computation.}
}
\value{
Returns a numeric vector for i-th cluster meancurve hausdorff distance from other clusters meancurves.
}
\description{
Calculates the Hausdorff distance between the i-th cluster meancurve and the other clusters meancurves.
}
