% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Betweenness.R
\name{Betweenness}
\alias{Betweenness}
\title{Betweenness}
\usage{
Betweenness(ClustCurve, MeanCurves)
}
\arguments{
\item{ClustCurve}{Data frame with five arguments : time, growth values, ID, cluster membership and feature values for each sample.}

\item{MeanCurves}{Matrix with the meancurve values on the columns according to different clusters.}
}
\value{
Returns a list with three arguments: (i) Betweenness, a matrix composed by two columns, the minimum cluster distance and the cluster name that achieves the minimum, (ii) CentroidDist, a K x K matrix (where K is the cluster number) containing centroids cluster distance among each cluster couple, (iii) Classes, a vector containing the cluster name membership for each dataset sample.
}
\description{
Computes the Betweenness across all clusters.
}
