% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_utils.R
\name{change_to_new_md}
\alias{change_to_new_md}
\title{Update metadata field}
\usage{
change_to_new_md(old_metadata, new_metadata)
}
\arguments{
\item{old_metadata}{\link{list} a list of element to be replace}

\item{new_metadata}{\link{list} a list of element to replace old's}
}
\value{
\link{list} a updated list with new metadata
}
\description{
Update metadata field
}
