% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circosplot.R
\name{circo.generation}
\alias{circo.generation}
\title{Circo plot of the connector clustering}
\usage{
circo.generation(dataFrom, dataTo, nameFrom, nameTo, filename)
}
\arguments{
\item{dataFrom}{starting dataframe with one column ID and the second one the connector cluster membership}

\item{dataTo}{ending dataframe with one column ID and the second one the connector cluster membership}

\item{nameFrom}{name of the experiment to print above the starting cluster names}

\item{nameTo}{name of the experiment to print above the ending cluster names}
}
\value{

}
\description{
Generates the circos plot given two connector clustering
}
\examples{

}
\author{
Cordero Francesca, Pernice Simone, Sirovich Roberta
}
