% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.symbol.R
\name{cluster.symbol}
\alias{cluster.symbol}
\title{Cluster Symbols}
\usage{
cluster.symbol(K)
}
\arguments{
\item{K}{Number of clusters.}
}
\value{
Returns a vector of K symbols.
}
\description{
Assigns a symbol for each K clusters.
}
