% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_choice.R
\name{ClusterChoice}
\alias{ClusterChoice}
\title{Cluster Choice}
\usage{
ClusterChoice(data, k, h = 1, p = 5, PCAperc = NULL)
}
\arguments{
\item{data}{CONNECTORList.}

\item{k}{Number of clusters.}

\item{h}{Dimension of the cluster mean space. If NULL, ClusterChoice set the h value equals to the number of PCA components needed to explain the 95 perc. of variability of the data.}

\item{PCAperc}{The PCA percentages calculated with the function PCAbarplot, if NULL then it must be inserted in input a value for h.
 @return
The matrices of the AIC and BIC values, a list of FCMList objects belonging to class funcyOutList for each h and k and  the plot generated using the Elbow Method.}
}
\description{
Calculates the BIC and AIC values considering k number of clusters and h dimension of the cluster mean space, and generates the plot based on the Elbow Method.
}
\examples{

GrowDataFile<-"data/1864dataset.xls"
AnnotationFile <-"data/1864info.txt"

CONNECTORList <- DataImport(GrowDataFile,AnnotationFile)

CONNECTORList<- DataTruncation(CONNECTORList,"Progeny",truncTime=60,labels = c("time","volume","Tumor Growth"))

# Specifying the h value
CONNECTORList.FCM <- ClusterChoice(CONNECTORList,k=c(2:6),h=2)

# Using the PCA percentaes instead of the h value.

perc <- PCA.Analysis(CONNECTORList$Dataset)$perc

CONNECTORList.FCM <- ClusterChoice(CONNECTORList,k=c(2:6),PCAperc=perc)

}
