% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{clustering}
\alias{clustering}
\title{Clustering}
\usage{
clustering(data, k, model)
}
\arguments{
\item{data}{CONNECTORList.}

\item{k}{Number of clusters.}

\item{model}{Model name, i.e. Malthus, Gompertz and Logistic.}
}
\value{
A list containing per each curve (i) the parameters of the fitting, (ii) the centers, (iii) the cluster of affinity and (iv) the mean values.
}
\description{
Fitting the cancer growth data with respect to the following growth function: Malthus, Gompertz and Logistic. The cancer growth fitted functions are then clustered.
}
\examples{

}
