% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi_methods.R
\name{cnt_disconnect.connector_dbi}
\alias{cnt_disconnect.connector_dbi}
\title{Disconnect from the database}
\usage{
\method{cnt_disconnect}{connector_dbi}(connector_object, ...)
}
\arguments{
\item{connector_object}{connector_dbi object}

\item{...}{Additional arguments passed to \link[DBI:dbDisconnect]{DBI::dbDisconnect}}
}
\description{
Disconnect from the database
}
\examples{
connector <- connector_dbi$new(RSQLite::SQLite())
connector$cnt_disconnect()
}
