% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_generics.R
\name{cnt_disconnect}
\alias{cnt_disconnect}
\title{Disconnect from the database}
\usage{
cnt_disconnect(connector_object, ...)
}
\arguments{
\item{connector_object}{\link{connector} The connector object to use.}

\item{...}{Additional arguments passed to the method for the individual connector.}
}
\value{
\link{invisible} \code{connector_object}
}
\description{
Disconnect from the database
}
