% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_generics.R
\name{cnt_download}
\alias{cnt_download}
\title{Download content from the connector}
\usage{
cnt_download(connector_object, name, file = basename(name), ...)
}
\arguments{
\item{connector_object}{\link{connector} The connector object to use.}

\item{name}{\link{character} Name of the content to read, write, or remove. Typically the table name.}

\item{file}{\link{character} Path to the file to download to or upload from}

\item{...}{Additional arguments passed to the method for the individual connector.}
}
\value{
\link{invisible} \code{file}
}
\description{
Download content from the connector
}
