% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_methods.R
\name{cnt_read.connector_fs}
\alias{cnt_read.connector_fs}
\title{Read files for the connector_fs class}
\usage{
\method{cnt_read}{connector_fs}(connector_object, name, ...)
}
\arguments{
\item{connector_object}{A connector_fs object}

\item{name}{Name of the file to read}

\item{...}{Other parameters to pass to the read_file function (depends on the extension of a file)}
}
\description{
Read files for the connector_fs class
}
\examples{
connector <- connector_fs$new(tempdir())
connector$write(iris, "iris.csv")
connector$read("iris.csv")
connector$remove("iris.csv")
}
