% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_generics.R, R/fs_methods.R
\name{cnt_read}
\alias{cnt_read}
\alias{cnt_read.connector_fs}
\title{Read content from the connector}
\usage{
cnt_read(connector_object, ...)

\method{cnt_read}{connector_fs}(connector_object, name, ...)
}
\arguments{
\item{connector_object}{\link{connector} The connector object to use.}

\item{...}{Additional arguments passed to the method for the individual connector.}
}
\value{
R object with the content. For rectangular data a \link{data.frame}.
}
\description{
Generic implementing how to read content from the different connector objects:

\itemize{
\item \link{connector_fs}: Uses \link{read_file} to read a given file.
The underlying function used, and thereby also the arguments available
through \code{...} depends on the file extension.
}
}
\examples{
# Write and read a CSV file using the file storage connector
cnt_fs <- connector_fs$new(tempdir())

cnt_fs |>
  cnt_write(iris, "iris.csv")

cnt_fs |>
  cnt_read("iris.csv") |>
  head()

}
