% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_methods.R
\name{cnt_remove.connector_fs}
\alias{cnt_remove.connector_fs}
\title{Remove a file or directory}
\usage{
\method{cnt_remove}{connector_fs}(connector_object, ...)
}
\arguments{
\item{connector_object}{connector_fs object}

\item{...}{file name in case of connector_fs}
}
\description{
Remove a file or directory
}
\examples{
connector <- connector_fs$new(tempdir())
connector$write(iris, "iris.csv")
connector$remove("iris.csv")
}
