% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_generics.R
\name{cnt_remove}
\alias{cnt_remove}
\title{Remove content from the connector}
\usage{
cnt_remove(connector_object, name, ...)
}
\arguments{
\item{connector_object}{\link{connector} The connector object to use.}

\item{name}{\link{character} Name of the content to read, write, or remove. Typically the table name.}

\item{...}{Additional arguments passed to the method for the individual connector.}
}
\value{
\link{invisible} \code{connector_object}
}
\description{
Remove content from the connector
}
