% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_generics.R, R/fs_methods.R
\name{cnt_remove_directory}
\alias{cnt_remove_directory}
\alias{cnt_remove_directory.connector_fs}
\title{Remove a directory}
\usage{
cnt_remove_directory(connector_object, name, ...)

\method{cnt_remove_directory}{connector_fs}(connector_object, name, ...)
}
\arguments{
\item{name}{\link{character} The name of the directory to remove}

\item{...}{Additional arguments passed to the method for the individual connector.}
}
\value{
\link{invisible} connector_object.
}
\description{
Generic implementing how to remove a directory for a connector.
Mostly relevant for file storage connectors.

\itemize{
\item \link{connector_fs}: Uses \link{unlink} with \code{recursive = TRUE} to remove a directory at the path of the connector.
}
}
\examples{
# Remove a directory from a file storage

cnt_fs <- connector_fs$new(tempdir())

cnt_fs |>
  cnt_create_directory("new_folder") |>
  cnt_list_content(pattern = "new_folder")

cnt_fs |>
  cnt_remove_directory("new_folder") |>
  cnt_list_content(pattern = "new_folder")

}
