% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_generics.R
\name{cnt_tbl}
\alias{cnt_tbl}
\title{Use dplyr verbs to interact with the remote database table}
\usage{
cnt_tbl(connector_object, name, ...)
}
\arguments{
\item{connector_object}{\link{connector} The connector object to use.}

\item{name}{\link{character} Name of the content to read, write, or remove. Typically the table name.}
}
\value{
A \link[dplyr:tbl]{dplyr::tbl} object.
}
\description{
Use dplyr verbs to interact with the remote database table
}
