% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_generics.R, R/fs_methods.R
\name{cnt_upload}
\alias{cnt_upload}
\alias{cnt_upload.connector_fs}
\title{Upload content to the connector}
\usage{
cnt_upload(connector_object, file, name = basename(file), ...)

\method{cnt_upload}{connector_fs}(connector_object, file, name = basename(file), ...)
}
\arguments{
\item{connector_object}{\link{connector} The connector object to use.}

\item{file}{\link{character} Path to the file to download to or upload from}

\item{name}{\link{character} Name of the content to read, write, or remove. Typically the table name.}

\item{...}{Additional arguments passed to the method for the individual connector.}
}
\value{
\link{invisible} connector_object.
}
\description{
Generic implementing of how to upload files to a connector:

\itemize{
\item \link{connector_fs}: Uses \code{\link[=file.copy]{file.copy()}} to copy the \code{file} to the file storage.
}
}
\examples{
# Upload file to a file storage

writeLines("this is an example", "example.txt")

cnt <- connector_fs$new(tempdir())

cnt |>
  cnt_list_content(pattern = "example.txt")

cnt |>
  cnt_upload("example.txt")

cnt |>
  cnt_list_content(pattern = "example.txt")

}
