% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi_methods.R
\name{cnt_write.connector_dbi}
\alias{cnt_write.connector_dbi}
\title{Write a table to the database}
\usage{
\method{cnt_write}{connector_dbi}(connector_object, x, name, ...)
}
\arguments{
\item{connector_object}{connector_dbi object}

\item{x}{Table to write}

\item{name}{Table name}

\item{...}{Additional arguments passed to \link[DBI:dbWriteTable]{DBI::dbWriteTable}}
}
\description{
Write a table to the database
}
\examples{
connector <- connector_dbi$new(RSQLite::SQLite())
connector$cnt_write(iris, "iris")
}
