% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_methods.R
\name{cnt_write.Connector_fs}
\alias{cnt_write.Connector_fs}
\title{Write a file based on this extension}
\usage{
\method{cnt_write}{Connector_fs}(connector_object, x, file, ...)
}
\arguments{
\item{connector_object}{A Connector_fs object}

\item{x}{Object to write}

\item{file}{Path to write the file}

\item{...}{Other parameters for write's functions}
}
\description{
Write a file based on this extension
}
\examples{
connector <- Connector_fs$new(tempdir())
connector$write(iris, "iris.csv")
connector$remove("iris.csv")

}
