% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector.R
\name{connect}
\alias{connect}
\title{Connect to a database}
\usage{
connect(yaml = NULL, ...)
}
\arguments{
\item{yaml}{A yaml file with the connection information}

\item{...}{Arguments to pass to the connector}
}
\value{
A new connector object
}
\description{
Connect to a database
}
