% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.R
\name{connect_dbi}
\alias{connect_dbi}
\title{Create DBI connector}
\usage{
connect_dbi(drv, ..., extra_class = NULL)
}
\arguments{
\item{drv}{DBI driver. See \link[DBI:dbConnect]{DBI::dbConnect} for details}

\item{...}{Additional arguments passed to \link[DBI:dbConnect]{DBI::dbConnect}}

\item{extra_class}{\link{character} Extra class added to the object. See details.}
}
\value{
A new \link{connector_dbi} object
}
\description{
Create a new DBI connector object. See \link{connector_dbi} for details.
}
\details{
The \code{extra_class} parameter allows you to create a subclass of the \code{connector_dbi} object.
This can be useful if you want to create a custom connection object for easier dispatch of new s3 methods,
while still inheriting the methods from the \code{connector_dbi} object.
}
\examples{
# Connect to in memory SQLite database

db <- connect_dbi(RSQLite::SQLite(), ":memory:")

db

# Create subclass connection

db_subclass <- connect_dbi(RSQLite::SQLite(), ":memory:", extra_class = "subclass")

db_subclass
class(db_subclass)

}
