% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector.R
\name{connect_from_yaml}
\alias{connect_from_yaml}
\title{Connect datasources to the connections from the yaml content}
\usage{
connect_from_yaml(yaml_content)
}
\arguments{
\item{yaml_content}{The yaml content}
}
\value{
A Connector object
}
\description{
Connect datasources to the connections from the yaml content
}
\examples{
yaml_file <- system.file("config", "default_config.yml", package = "connector")
yaml_content <- yaml::read_yaml(yaml_file, eval.expr = TRUE)
# create the connections
connect <- connect_from_yaml(yaml_content)
}
