% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_logger_generics.R
\name{read_cnt.ConnectorLogger}
\alias{read_cnt.ConnectorLogger}
\alias{tbl_cnt.ConnectorLogger}
\alias{connector-logger-methods}
\alias{write_cnt.ConnectorLogger}
\alias{remove_cnt.ConnectorLogger}
\alias{list_content_cnt.ConnectorLogger}
\alias{upload_cnt.ConnectorLogger}
\alias{download_cnt.ConnectorLogger}
\alias{print.ConnectorLogger}
\title{Log Read Operation for ConnectorLogger class}
\usage{
\method{read_cnt}{ConnectorLogger}(connector_object, name, ...)

\method{tbl_cnt}{ConnectorLogger}(connector_object, name, ...)

\method{write_cnt}{ConnectorLogger}(connector_object, x, name, ...)

\method{remove_cnt}{ConnectorLogger}(connector_object, name, ...)

\method{list_content_cnt}{ConnectorLogger}(connector_object, ...)

\method{upload_cnt}{ConnectorLogger}(
  connector_object,
  file,
  name = basename(file),
  overwrite = zephyr::get_option("overwrite", "connector"),
  ...
)

\method{download_cnt}{ConnectorLogger}(connector_object, name, file = basename(name), ...)

\method{print}{ConnectorLogger}(x, ...)
}
\arguments{
\item{connector_object}{The ConnectorLogger object.}

\item{name}{The name of the file in the connector.}

\item{...}{Additional arguments}

\item{x}{The connector logger object}

\item{file}{The local file path to download to.}

\item{overwrite}{Overwrite existing content if it exists in the connector?
See \link{connector-options} for details. Default can be set globally with
\code{options(connector.overwrite = TRUE/FALSE)} or environment variable
\code{R_CONNECTOR_OVERWRITE}.. Default: \code{FALSE}.}
}
\value{
The result of the read operation.

Invisible result of the write operation.

The result of the remove operation.

The result of the read operation.

Invisible result of the upload operation.

The result of the download operation.

The result of the print operation
}
\description{
Implementation of the log_read_connector function for the ConnectorLogger
class.

Implementation of the log_write_connector function for the ConnectorLogger
class.

Implementation of the log_remove_connector function for the ConnectorLogger
class.

Implementation of the log_read_connector function for the ConnectorLogger
class.

Implementation of the upload_cnt function for the ConnectorLogger
class.

Implementation of the download_cnt function for the ConnectorLogger
class.

This function prints the connector logger.
}
