% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector-options.R
\name{connector-options}
\alias{connector-options}
\title{Options for connector}
\description{
\subsection{verbosity_level}{

Verbosity level for functions in connector.
See \link[zephyr:verbosity_level]{zephyr::verbosity_level} for details.
\itemize{
\item Default: \code{"verbose"}
\item Option: \code{connector.verbosity_level}
\item Environment: \code{R_CONNECTOR_VERBOSITY_LEVEL}
}
}

\subsection{overwrite}{

Overwrite existing content if it exists in the connector?
See \link{connector-options} for details. Default can be set globally with
\code{options(connector.overwrite = TRUE/FALSE)} or environment variable
\code{R_CONNECTOR_OVERWRITE}.
\itemize{
\item Default: \code{FALSE}
\item Option: \code{connector.overwrite}
\item Environment: \code{R_CONNECTOR_OVERWRITE}
}
}

\subsection{logging}{

Add logging capability to connectors using \code{\link[=add_logs]{add_logs()}}.
When \code{TRUE}, all connector operations will be logged to the console and
to whirl log HTML files. See \link{log-functions} for available
logging functions.
\itemize{
\item Default: \code{FALSE}
\item Option: \code{connector.logging}
\item Environment: \code{R_CONNECTOR_LOGGING}
}
}

\subsection{default_ext}{

Default extension to use when reading and writing files when not
specified in the file name. E.g. with the default 'csv', files are assumed
to be in CSV format if not specified.
\itemize{
\item Default: \code{"csv"}
\item Option: \code{connector.default_ext}
\item Environment: \code{R_CONNECTOR_DEFAULT_EXT}
}
}
}
