% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector-package.R
\docType{package}
\name{connector-package}
\alias{connector-package}
\title{connector: Streamlining Data Access in Clinical Research}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a consistent interface for connecting R to various data sources including file systems and databases. Designed for clinical research, 'connector' streamlines access to 'ADAM', 'SDTM' for example. It helps to deal with multiple data formats through a standardized API and centralized configuration.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/novonordisk-opensource/connector/}
  \item \url{https://github.com/NovoNordisk-OpenSource/connector}
  \item Report bugs at \url{https://github.com/NovoNordisk-OpenSource/connector/issues}
}

}
\author{
\strong{Maintainer}: Cervan Girard \email{cgid@novonordisk.com}

Authors:
\itemize{
  \item Aksel Thomsen \email{oath@novonordisk.com}
  \item Vladimir Obucina \email{vlob@novonordisk.com}
}

Other contributors:
\itemize{
  \item Novo Nordisk A/S [copyright holder]
}

}
\keyword{internal}
