% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector.R
\name{connector_}
\alias{connector_}
\title{R6 general connector class}
\description{
R6 general connector class

R6 general connector class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-connector_-clone}{\code{connector_$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connector_-clone"></a>}}
\if{latex}{\out{\hypertarget{method-connector_-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{connector_$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
