% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.R
\name{connector_dbi_methods}
\alias{connector_dbi_methods}
\alias{cnt_disconnect}
\alias{cnt_tbl}
\title{disconnect method for connector object}
\usage{
cnt_disconnect(connector_object, ...)

cnt_tbl(connector_object, ...)
}
\description{
disconnect method for connector object

tbl method for connector object
}
