% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{connector_fs}
\alias{connector_fs}
\title{Class connector_fs}
\description{
The connector_fs class is a file system connector for accessing and manipulating files in a local file system.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-connector_fs-new}{\code{connector_fs$new()}}
\item \href{#method-connector_fs-list_content}{\code{connector_fs$list_content()}}
\item \href{#method-connector_fs-construct_path}{\code{connector_fs$construct_path()}}
\item \href{#method-connector_fs-read}{\code{connector_fs$read()}}
\item \href{#method-connector_fs-write}{\code{connector_fs$write()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connector_fs-new"></a>}}
\if{latex}{\out{\hypertarget{method-connector_fs-new}{}}}
\subsection{Method \code{new()}}{
Initializes the connector_fs class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{connector_fs$new(path, access = "rw")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path to the file system}

\item{\code{access}}{Access type ("rw" by default)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connector_fs-list_content"></a>}}
\if{latex}{\out{\hypertarget{method-connector_fs-list_content}{}}}
\subsection{Method \code{list_content()}}{
Returns the list of files in the specified path
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{connector_fs$list_content(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Other parameters to pass to the list.files function}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connector_fs-construct_path"></a>}}
\if{latex}{\out{\hypertarget{method-connector_fs-construct_path}{}}}
\subsection{Method \code{construct_path()}}{
Constructs a complete path by combining the specified access path with the provided elements
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{connector_fs$construct_path(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Elements to construct the path}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connector_fs-read"></a>}}
\if{latex}{\out{\hypertarget{method-connector_fs-read}{}}}
\subsection{Method \code{read()}}{
Reads the content of the specified file using the private access path and additional options
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{connector_fs$read(name, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the file to read}

\item{\code{...}}{Other parameters to pass to the read_file function (depends on the extension of a file)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connector_fs-write"></a>}}
\if{latex}{\out{\hypertarget{method-connector_fs-write}{}}}
\subsection{Method \code{write()}}{
Writes the specified content to the specified file using the private access path and additional options
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{connector_fs$write(x, file, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Content to write to the file}

\item{\code{file}}{File name}

\item{\code{...}}{Other parameters to pass to the write_file function (depends on the extension of a file)}
}
\if{html}{\out{</div>}}
}
}
}
