% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{connector_fs_methods}
\alias{connector_fs_methods}
\alias{cnt_download}
\alias{cnt_upload}
\alias{cnt_create_directory}
\alias{cnt_remove_directory}
\title{Additional methods for file storage connectors}
\usage{
cnt_download(connector_object, ...)

cnt_upload(connector_object, ...)

cnt_create_directory(connector_object, ...)

cnt_remove_directory(connector_object, ...)
}
\arguments{
\item{connector_object}{A \link{connector_fs} object to be able to use functions from it}

\item{...}{Additional arguments passed to the methods}
}
\description{
These methods are additional S3 methods for  \link{connector_fs}.
}
\seealso{
\link{connector_methods}
}
