% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{Connector_fs_object}
\alias{Connector_fs_object}
\alias{Connector_fs}
\title{Connector for file system}
\description{
The connector_fs class is a file system connector for accessing and manipulating files in a local file system.
}
\section{Super class}{
\code{\link[connector:Connector]{connector::Connector}} -> \code{Connector_fs}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Connector_fs-new}{\code{Connector_fs$new()}}
\item \href{#method-Connector_fs-construct_path}{\code{Connector_fs$construct_path()}}
\item \href{#method-Connector_fs-get_path}{\code{Connector_fs$get_path()}}
\item \href{#method-Connector_fs-clone}{\code{Connector_fs$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="connector" data-topic="Connector" data-id="list_content"><a href='../../connector/html/Connector.html#method-Connector-list_content'><code>connector::Connector$list_content()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connector" data-topic="Connector" data-id="read"><a href='../../connector/html/Connector.html#method-Connector-read'><code>connector::Connector$read()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connector" data-topic="Connector" data-id="remove"><a href='../../connector/html/Connector.html#method-Connector-remove'><code>connector::Connector$remove()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connector" data-topic="Connector" data-id="write"><a href='../../connector/html/Connector.html#method-Connector-write'><code>connector::Connector$write()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connector_fs-new"></a>}}
\if{latex}{\out{\hypertarget{method-Connector_fs-new}{}}}
\subsection{Method \code{new()}}{
Initializes the connector_fs class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connector_fs$new(path, access = "rw")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path to the file system}

\item{\code{access}}{Access type ("rw" by default)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connector_fs-construct_path"></a>}}
\if{latex}{\out{\hypertarget{method-Connector_fs-construct_path}{}}}
\subsection{Method \code{construct_path()}}{
Constructs a complete path by combining the specified access path with the provided elements
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connector_fs$construct_path(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Elements to construct the path}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connector_fs-get_path"></a>}}
\if{latex}{\out{\hypertarget{method-Connector_fs-get_path}{}}}
\subsection{Method \code{get_path()}}{
Returns the path of the file system
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connector_fs$get_path()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Path to the file system
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connector_fs-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Connector_fs-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connector_fs$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
