% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector.R
\name{connector_methods}
\alias{connector_methods}
\alias{cnt_read}
\alias{cnt_write}
\alias{cnt_remove}
\alias{cnt_list_content}
\title{Defaults methods for all connector object}
\usage{
cnt_read(connector_object, ...)

cnt_write(connector_object, ...)

cnt_remove(connector_object, ...)

cnt_list_content(connector_object, ...)
}
\arguments{
\item{connector_object}{A connector object to be able to use functions from it}

\item{...}{Additional arguments passed to the method}
}
\value{
The result of the read method

The result of the write method

The result of the remove method

The result of the list_content method
}
\description{
These methods are a S3 method that dispatches to the specific methods for the connector object.
}
\details{
For example, if you have a connector object \code{connector_fs}, you can use \code{cnt_read(connector_fs, "file.csv")} to read the file. It will be dispatch to the \code{cnt_read.connector_fs} method.
Why? The main aim is to allow the user to use the same function for different connector objects by using the builder function \code{connector_fs} and add an "extra_class".
By doing so, you can create a subclass of the \code{connector_fs} object and dispatch to the specific methods for this subclass.
For example, if you have a subclass \code{subclass}, you can use \code{cnr_read(subclass, "file.csv")} to read the file. It will be dispatch to the \code{cnt_read.subclass} method. And you can still use the \code{connector_fs} methods.
}
