% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector.R
\name{Connector_object}
\alias{Connector_object}
\alias{Connector}
\title{R6 class for a general Connector object}
\description{
R6 class for a general Connector object

R6 class for a general Connector object
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Connector-list_content}{\code{Connector$list_content()}}
\item \href{#method-Connector-read}{\code{Connector$read()}}
\item \href{#method-Connector-write}{\code{Connector$write()}}
\item \href{#method-Connector-remove}{\code{Connector$remove()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connector-list_content"></a>}}
\if{latex}{\out{\hypertarget{method-Connector-list_content}{}}}
\subsection{Method \code{list_content()}}{
List tables in the database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connector$list_content(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments passed to \link[DBI:dbListTables]{DBI::dbListTables}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \link{character} vector of table names
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connector-read"></a>}}
\if{latex}{\out{\hypertarget{method-Connector-read}{}}}
\subsection{Method \code{read()}}{
Read a table from the database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connector$read(name, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{\link{character} Table name}

\item{\code{...}}{Additional arguments passed to \link[DBI:dbReadTable]{DBI::dbReadTable}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \link{data.frame}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connector-write"></a>}}
\if{latex}{\out{\hypertarget{method-Connector-write}{}}}
\subsection{Method \code{write()}}{
Write a table to the database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connector$write(x, name, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\link{data.frame} Table to write}

\item{\code{name}}{\link{character} Table name}

\item{\code{...}}{Additional arguments passed to \link[DBI:dbWriteTable]{DBI::dbWriteTable}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connector-remove"></a>}}
\if{latex}{\out{\hypertarget{method-Connector-remove}{}}}
\subsection{Method \code{remove()}}{
Remove a table from the database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connector$remove(name, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{\link{character} Table name}

\item{\code{...}}{Additional arguments passed to \link[DBI:dbRemoveTable]{DBI::dbRemoveTable}}
}
\if{html}{\out{</div>}}
}
}
}
