% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger_generics.R
\docType{data}
\name{ConnectorLogger}
\alias{ConnectorLogger}
\alias{print.ConnectorLogger}
\title{Create a New Connector Logger}
\format{
An object of class \code{ConnectorLogger} of length 0.
}
\usage{
ConnectorLogger

\method{print}{ConnectorLogger}(x, ...)
}
\arguments{
\item{x}{object to print}

\item{...}{parameters passed to the print method}
}
\value{
An S3 object of class "ConnectorLogger" containing:
\itemize{
\item An empty list
\item Class attribute set to "ConnectorLogger"
}
}
\description{
Creates a new empty connector logger object of class
"ConnectorLogger".
This is an internal utility class that initializes a logging structure for
connector operations. Logs are added to connectors using \code{\link[=add_logs]{add_logs()}}.
}
\details{
Create a New Connector Logger
}
\examples{
logger <- ConnectorLogger
class(logger) # Returns "ConnectorLogger"
str(logger) # Shows empty list with class attribute

}
\keyword{datasets}
