% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectors.R
\name{Connectors}
\alias{Connectors}
\title{A Connectors object, a special list with R6 objects.}
\usage{
Connectors(...)
}
\arguments{
\item{...}{Arguments to pass to the connector}
}
\description{
A Connectors object, a special list with R6 objects.
}
