% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conts_datasources.R
\name{connectors_to_datasources}
\alias{connectors_to_datasources}
\title{Transform Test Data to Datasources}
\usage{
connectors_to_datasources(data)
}
\arguments{
\item{data}{A list of function calls as expressions.}
}
\value{
A list with a 'datasources' element containing the transformed backends.
}
\description{
This function takes a list of function calls, extracts their information,
transforms them into backends, and finally wraps them in a datasources structure.
}
