% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_backend.R
\name{create_backend}
\alias{create_backend}
\title{Create a backend}
\usage{
create_backend(backend)
}
\arguments{
\item{backend}{The backend to create}
}
\value{
A new backend based on R6 class
}
\description{
Create a backend
}
\examples{
yaml_file <- system.file("config", "example_for_generic.yml", package = "connector")
yaml_content <- read_yaml_config(yaml_file)

only_one <- yaml_content[["connections"]][[1]][["backend"]]

test <- create_backend(only_one)

}
