% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi_backend_tools.R
\name{create_backend_dbi}
\alias{create_backend_dbi}
\title{Create a backend for the DBI (databases)}
\usage{
create_backend_dbi(backend)
}
\arguments{
\item{backend}{The backend to create and because of DBI, a "drv" is mandatory}
}
\value{
A new backend based on R6 class
}
\description{
Create a backend for the DBI (databases)
}
\examples{
yaml_file <- system.file("config", "default_config.yml", package = "connector")
yaml_content <- read_yaml_config(yaml_file)

only_one <- yaml_content[["connections"]][[2]][["backend"]]

test <- create_backend_dbi(only_one)

}
