% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_backend_tools.R
\name{create_backend_fs}
\alias{create_backend_fs}
\title{Create a backend for the file system}
\usage{
create_backend_fs(backend)
}
\arguments{
\item{backend}{The backend to create and because of file system, a "path" is mandatory}
}
\value{
A new backend based on R6 class
}
\description{
Create a backend for the file system
}
\examples{
yaml_file <- system.file("config", "default_config.yml", package = "connector")
yaml_content <- yaml::read_yaml(yaml_file, eval.expr = TRUE)
only_one <- extract_connections(yaml_content)[[1]]
## Extract fct
my_backend <- only_one \%>\%
  extract_backends()
name <- only_one \%>\%
  extract_con()
# Create the backend
test <- create_backend_fs(yaml_content = yaml_content, backend = my_backend, name = name)
}
