% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector.R
\name{create_connection}
\alias{create_connection}
\title{Create a connection object depending on the backend type}
\usage{
create_connection(yaml_content, connection)
}
\arguments{
\item{yaml_content}{The yaml content}

\item{connection}{The connection object from the yaml}
}
\description{
Create a connection object depending on the backend type
}
