% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_backend_tools.R
\name{custom_path_or_not}
\alias{custom_path_or_not}
\title{Check if the path is a custom path to pass to glue}
\usage{
custom_path_or_not(path)
}
\arguments{
\item{path}{The path to check}
}
\value{
A boolean
}
\description{
Check if the path is a custom path to pass to glue
}
