% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFrameWithinness_i.R
\name{DataFrameWithinness.i}
\alias{DataFrameWithinness.i}
\title{Withinness data frame}
\usage{
DataFrameWithinness.i(ClustCurve, MeanCurves, i, ClustSymbol, shift = 0)
}
\arguments{
\item{ClustCurve}{Data frame with five arguments : time, growth values, ID, cluster membership and feature values for each sample.}

\item{MeanCurves}{Matrix with the meancurve values on the columns according to different clusters.}

\item{i}{Value included from 0 to K (number of clusters) representing the cluster involved in withinness computation.}

\item{ClustSymbol}{Vector of the cluster symbols.}

\item{shift}{Vector representing the distances between the first withinness circle center and the other centers.}
}
\value{
Returns a list composed by two data frames: (i) "circles" containing coordinates and radius for i-th cluster withinness circles measures, (ii) "WithDist" containing the i-th cluster curves withinness distance from i-th cluster centroid/meancurve.
}
\description{
Creates a dataframe for i-th cluster withinness measures.
}
