% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataImport.R
\name{DataImport}
\alias{DataImport}
\title{Data Import}
\usage{
DataImport(file1, file2)
}
\arguments{
\item{file1}{the name of the excel file which the data are to be read from. The growth data associated with an experiment must be reported using a pair of columns. The first column must contains the time points, suggested column header "Time". The second column must contains the data volume, the column header is the sample name.}

\item{file2}{the name of a cvs file which the annontation data are reported. Each row of the file containes: the identifier (ID) of the sample, e.g. an integer number, the sample name used in file1 and a list of features associated with the sample.}
}
\value{
List with four arguments: (i) data frame reporting three variables (ID, data and time values), (ii) the vector reporting the number of observations collected per sample, (iii) the data frame with curves labeled according to target file features and (iv) the vector for overall time grid.
}
\description{
Reads the files in a table format and create a list storing all the information.
}
\examples{

GrowDataFile<-"data/1864dataset.xls"
AnnotationFile <-"data/1864info.txt"

CONNECTORList <- DataImport(GrowDataFile,AnnotationFile)

}
