% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataTrunc.R
\name{DataTrunc}
\alias{DataTrunc}
\title{DataTrunc}
\usage{
DataTrunc(data, truncTime = NULL)
}
\arguments{
\item{data}{CONNECTORList.}

\item{truncTime}{An integer number corresponding to the time at which truncate the curves.}
}
\value{
The CONNECTORList updated with the following arguments: a data frame with three variables (ID curves, volume and time values truncated at the chosen time), a vector collecting the number of truncated observations collected per sample, a data frame with curves labeled according to target file feature chosen and a vector for overall truncated time grid.
}
\description{
Truncates the cancer growth data at a specific time point chosen by the user.
}
\examples{

GrowDataFile<-"data/1864dataset.xls"
AnnotationFile <-"data/1864info.txt"

CONNECTORList <- DataImport(GrowDataFile,AnnotationFile)

CONNECTORList<- DataTrunc(CONNECTORList,truncTime=60)

}
