% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataTruncation.R
\name{DataTruncation}
\alias{DataTruncation}
\title{DataTruncation}
\usage{
DataTruncation(data, feature, truncTime = NULL, labels = NULL,
  save = FALSE, path = NULL)
}
\arguments{
\item{data}{CONNECTORList.}

\item{feature}{The column name reported in the AnnotationFile containing the feature interesting for the user to be investigated.}

\item{truncTime}{An integer number corresponding to the time at which truncate the curves.}

\item{labels}{Vector containing the text for the title of axis and plot title.}

\item{save}{If TRUE the plot is saved in a pdf file.}

\item{path}{Path to save plot to (combined with file name). If it is missing, the plot is saved in the working directory.}
}
\value{
DataTruncation returns the line plot of growth curves with a vertical line at the truncation time and the CONNECTORList updated with the following arguments: a data frame with three variables (ID curves, volume and time values truncated at the chosen time), a vector collecting the number of truncated observations collected per sample, a data frame with curves labeled according to target file feature chosen and a vector for overall truncated time grid.
}
\description{
Truncates the cancer growth data at a specific time point chosen by the user.
}
\examples{

GrowDataFile<-"data/1864dataset.xls"
AnnotationFile <-"data/1864info.txt"

CONNECTORList <- DataImport(GrowDataFile,AnnotationFile)

CONNECTORList<- DataTruncation(CONNECTORList,"Progeny",truncTime=60,labels = c("time","volume","Tumor Growth"))

}
