% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataVisualization.R
\name{DataVisualization}
\alias{DataVisualization}
\title{Data Visualitation}
\usage{
DataVisualization(data, feature, labels = NULL, save = FALSE, path = NULL)
}
\arguments{
\item{data}{CONNECTORList.}

\item{feature}{The column name reported in the AnnotationFile containing the feature interesting for the user to be investigated.}

\item{labels}{Vector containing the text for the title of axis and plot title.}

\item{save}{If TRUE the plot is saved in a pdf file.}

\item{path}{Path to save plot to (combined with file name). If it is missing, the plot is saved in the working directory.}
}
\value{
A plot with the density time grid and the line plot of cancer growth data as a ggplot object.
}
\description{
Visualization of the time grid and the line plot of cancer growth data.
}
\examples{

GrowDataFile<-"data/1864dataset.xls"
AnnotationFile <-"data/1864info.txt"

CONNECTORList <- DataImport(GrowDataFile,AnnotationFile)

DataVisualization(CONNECTORList,"Progeny",labels = c("time","volume","Tumor Growth"))

}
