% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumberKnotsChoice.R
\name{DimensionBasis.Choice}
\alias{DimensionBasis.Choice}
\title{Choice of the B-spline dimension}
\usage{
DimensionBasis.Choice(data, p.values)
}
\arguments{
\item{data}{CONNECTORList. (see \code{\link{DataImport}})}

\item{p}{The vector of the dimension of the natural cubic spline basis.}
}
\value{
DimensionBasis.Choice returns a plot with ...as a ggplot object.
}
\description{

}
\examples{

GrowDataFile<-"data/1864dataset.xls"
AnnotationFile <-"data/1864info.txt"

CONNECTORList <- DataImport(GrowDataFile,AnnotationFile)

CONNECTORList<- DataTruncation(CONNECTORList,"Progeny",truncTime=60,labels = c("time","volume","Tumor Growth"))


### Calculation of k and fitting using FCM
# Specifying the h value

.................


}
\references{
Gareth M. James and Catherine A. Sugar, (2003). Clustering for Sparsely Sampled Functional Data. Journal of the American Statistical Association.
}
\seealso{
\code{\link[funcy]{funcit}}.
}
