% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi_methods.R
\name{disconnect.Connector_dbi}
\alias{disconnect.Connector_dbi}
\title{Disconnect from the database}
\usage{
\method{disconnect}{Connector_dbi}(connector_object, ...)
}
\arguments{
\item{connector_object}{Connector_dbi object}

\item{...}{Additional arguments passed to \link[DBI:dbDisconnect]{DBI::dbDisconnect}}
}
\description{
Disconnect from the database
}
\examples{
connector <- Connector_dbi$new(RSQLite::SQLite())
connector$disconnect()
}
