% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_run.R
\name{docker.run}
\alias{docker.run}
\title{Run docker container}
\usage{
docker.run()
}
\value{
0 if success, 1 if dockerid file is present, 2 if docker execution fails.
}
\description{
This is an internal function executing a docker container.
}
\examples{
\dontrun{
    #running runDocker
     docker.run(params=NULL)

}
}
\author{
Beccuti Marco, Pernice Simone
}
