% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector.R
\name{.connector}
\alias{.connector}
\title{Create a connector}
\usage{
.connector(extra_class = NULL)
}
\arguments{
\item{extra_class}{\link{character} Extra class to assign to the new connector.}
}
\value{
A new \link{Connector} object
}
\description{
Initializes the connector object.
See \link{Connector} for details.
}
\details{
The \code{extra_class} parameter allows you to create a subclass of the
\code{Connector} object. This can be useful if you want to create
a custom connection object for easier dispatch of new s3 methods, while still
inheriting the methods from the \code{Connector} object.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

# Create a connector object
cnt <- .connector()
cnt

# Create subclass connection
cnt_subclass <- .connector(
  extra_class = "subclass"
)
cnt_subclass
class(cnt_subclass)
\dontshow{\}) # examplesIf}
}
