% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_logger_generics.R
\name{download_cnt.ConnectorLogger}
\alias{download_cnt.ConnectorLogger}
\title{Download Operation for ConnectorLogger class}
\usage{
\method{download_cnt}{ConnectorLogger}(connector_object, name, file = basename(name), ...)
}
\arguments{
\item{connector_object}{The ConnectorLogger object.}

\item{name}{The name of the file in the connector.}

\item{file}{The local file path to download to.}

\item{...}{Additional parameters.}
}
\value{
The result of the download operation.
}
\description{
Implementation of the download_cnt function for the ConnectorLogger
class.
}
