% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_files.R
\name{error_extension}
\alias{error_extension}
\title{Error extension
Function to call when no method is found for the extension}
\usage{
error_extension()
}
\description{
Error extension
Function to call when no method is found for the extension
}
