% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_files.R
\name{example_read_ext}
\alias{example_read_ext}
\title{Example for creating a new method for reading files}
\usage{
example_read_ext()
}
\description{
Example for creating a new method for reading files
}
\examples{
example_read_ext()
}
