% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conts_datasources.R
\name{extract_and_process_params}
\alias{extract_and_process_params}
\title{Extract and Process Parameters}
\usage{
extract_and_process_params(expr, formal_args)
}
\arguments{
\item{expr}{The parsed expression of the function call.}

\item{formal_args}{The formal arguments of the function.}
}
\value{
A list of processed parameters.
}
\description{
Extracts parameters from the function call and processes them.
}
\keyword{internal}
