% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conts_datasources.R
\name{extract_base_info}
\alias{extract_base_info}
\title{Extract Base Information}
\usage{
extract_base_info(full_func_name, is_r6)
}
\arguments{
\item{full_func_name}{The full name of the function (potentially including package).}

\item{is_r6}{Boolean indicating whether it's an R6 class constructor.}
}
\value{
A list with package_name and func_name.
}
\description{
Extracts the package name and function/class name from the full function name.
}
\keyword{internal}
