% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_backend_tools.R
\name{extract_custom_path}
\alias{extract_custom_path}
\title{Extract custom path from metadata information if needed}
\usage{
extract_custom_path(yaml_content, path)
}
\arguments{
\item{yaml_content}{The yaml content}

\item{path}{The path to extract}
}
\value{
The custom path
}
\description{
Extract custom path from metadata information if needed
}
