% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_yaml.R
\name{extract_element}
\alias{extract_element}
\alias{extract_metadata}
\alias{extract_connections}
\alias{extract_datasources}
\alias{extract_backends}
\alias{extract_con}
\title{Extract an element from a yaml content}
\usage{
extract_element(yaml_content, element)

extract_metadata(yaml_content)

extract_connections(yaml_content)

extract_datasources(yaml_content)

extract_backends(yaml_content)

extract_con(yaml_content)
}
\arguments{
\item{yaml_content}{The yaml content}

\item{element}{The element to extract}
}
\value{
The element extracted

The metadata extracted
}
\description{
Extract metadata from a yaml file
}
\examples{
yaml_file <- system.file("config", "default_config.yml", package = "connector")
yaml_content <- yaml::read_yaml(yaml_file, eval.expr = TRUE)
extract_element(yaml_content, "metadata")
yaml_file <- system.file("config", "default_config.yml", package = "connector")
yaml_content <- yaml::read_yaml(yaml_file, eval.expr = TRUE)
extract_metadata(yaml_content)
}
