% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conts_datasources.R
\name{extract_function_info}
\alias{extract_function_info}
\title{Extract Function Information}
\usage{
extract_function_info(func_string)
}
\arguments{
\item{func_string}{A character string representing the function call.}
}
\value{
A list with class "clean_fct_info" containing:
\item{function_name}{The name of the function or R6 class}
\item{parameters}{A list of parameters passed to the function}
\item{is_r6}{A boolean indicating whether it's an R6 class constructor}
\item{package_name}{The name of the package containing the function}
}
\description{
This function extracts detailed information about a function call,
including its name, package, parameters, and whether it's an R6 class constructor.
}
