% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_backend_tools.R
\name{extract_info_from_metadata}
\alias{extract_info_from_metadata}
\title{Extract information from metadata}
\usage{
extract_info_from_metadata(extracted_metadata, path)
}
\arguments{
\item{extracted_metadata}{The metadata extracted from the yaml file}

\item{path}{The path to extract information from}
}
\value{
The information extracted
}
\description{
Extract information from metadata
}
