% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtrapolationFunctions.R
\name{ExtrapolationFuncs}
\alias{ExtrapolationFuncs}
\alias{BoxPlot.Extrapolation}
\alias{ConsMatrix.Extrapolation}
\alias{MostProbableClustering.Extrapolation}
\title{Extrapolate objects from StabilityAnalysis}
\usage{
BoxPlot.Extrapolation(stability.list, h)

ConsMatrix.Extrapolation(stability.list, h, G)

MostProbableClustering.Extrapolation(stability.list, h, G)
}
\arguments{
\item{stability.list}{CONNECTORList obtained from the StabilityAnalysis function. (see \code{\link{StabilityAnalysis}})}

\item{h}{The  number representing the dimension of the cluster mean space (see \code{\link{PCA.Analysis}}).}

\item{G}{The number of clusters.}
}
\description{
Extrapolate objects from StabilityAnalysis output list (see \code{\link{StabilityAnalysis}}).
}
