% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{ExtrapolationNewFuncs}
\alias{ExtrapolationNewFuncs}
\alias{BoxPlot.ExtrapolationNew}
\alias{ConsMatrix.ExtrapolationNew}
\alias{MostProbableClustering.ExtrapolationNew}
\title{Extrapolate objects from StabilityAnalysis}
\usage{
BoxPlot.ExtrapolationNew(stability.list)

ConsMatrix.ExtrapolationNew(stability.list, data)

MostProbableClustering.ExtrapolationNew(stability.list, G)
}
\arguments{
\item{stability.list}{The list obtained from the StabilityAnalysis function. (see \code{\link{StabilityAnalysis}})}

\item{G}{The number of clusters.}

\item{h}{The number representing the dimension of the cluster mean space (see \code{\link{PCA.Analysis}}).}
}
\description{
Extrapolate objects from StabilityAnalysis output list (see \code{\link{StabilityAnalysis}}).
}
\details{
\itemize{
\item{BoxPlot.Extrapolation}{ extrapolates from StabilityAnalysis output list the box plot fixing the h value.}
\item{ConsMatrix.Extrapolation}{extrapolates from StabilityAnalysis output list the Consensus Matrix for G and h fixed.}
\item{MostProbableClustering.Extrapolation}{extrapolates from StabilityAnalysis output list the most probable cluster among the several runs obtained  for G and h fixed.}
}
}
\author{
Cordero Francesca, Pernice Simone, Sirovich Roberta
}
