% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_files.R
\name{find_file}
\alias{find_file}
\title{Find File}
\usage{
find_file(name, root)
}
\arguments{
\item{name}{Name of a file}

\item{root}{Path to the root folder}
}
\value{
A full name path to the file or a error if multiples files or 0.
}
\description{
Find File
}
