% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCM_fclust.R
\name{fitfclust}
\alias{fitfclust}
\title{FCM functions}
\usage{
fitfclust(x = NULL, curve = NULL, timeindex = NULL, data = NULL,
  q = 5, h = 1, p = 5, K = 2, tol = 0.001, maxit = 20,
  pert = 0.01, grid = seq(0, 1, length = 100), hard = F, plot = F,
  trace = F, seed = NULL)
}
\description{
FCM functions
}
