% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_backend_tools.R
\name{get_backend_fct}
\alias{get_backend_fct}
\title{Get the backend function}
\usage{
get_backend_fct(backend_type)
}
\arguments{
\item{backend_type}{The type of the backend, by default it is connector_fs or connector_db}
}
\value{
The backend function
}
\description{
Get the backend function
}
\examples{
get_backend_fct("connector_fs")
}
