% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector.R
\name{get_connections}
\alias{get_connections}
\title{Create all connections from the yaml content}
\usage{
get_connections(yaml_content)
}
\arguments{
\item{yaml_content}{The yaml content}
}
\description{
Create all connections from the yaml content
}
