% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conts_datasources.R
\name{get_r6_specific_info}
\alias{get_r6_specific_info}
\title{Get R6 Class Specific Information}
\usage{
get_r6_specific_info(package_name, func_name)
}
\arguments{
\item{package_name}{The name of the package containing the R6 class.}

\item{func_name}{The name of the R6 class.}
}
\value{
A list with the initialize method and its formal arguments.
}
\description{
Retrieves the initialize method and its formal arguments for R6 classes.
}
\keyword{internal}
