% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conts_datasources.R
\name{get_standard_specific_info}
\alias{get_standard_specific_info}
\title{Get Standard Function Specific Information}
\usage{
get_standard_specific_info(package_name, func_name)
}
\arguments{
\item{package_name}{The name of the package containing the function.}

\item{func_name}{The name of the function.}
}
\value{
A list with the function object and its formal arguments.
}
\description{
Retrieves the function object and its formal arguments for standard functions.
}
\keyword{internal}
