% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrowthCurve.R
\name{GrowthCurve}
\alias{GrowthCurve}
\title{Growth curves}
\usage{
GrowthCurve(data, feature, labels = NULL)
}
\arguments{
\item{data}{CONNECTORList.}

\item{feature}{The column name reported in the AnnotationFile containing the feature interesting for the user to be investigated.}

\item{labels}{Vector containing the text for the axis and plot title.}
}
\value{
A list containing the line plot and the CONNECTORList.
}
\description{
Visualization of the line plot of the cancer growth data.
The curves are colored with respect of the feature chosen by the user and reported in the AnnotationFile.
}
\examples{
GrowDataFile<-"data/1864dataset.xls"
AnnotationFile <-"data/1864info.txt"

CONNECTORList <- DataImport(GrowDataFile,AnnotationFile)

GrowthCurve(CONNECTORLis,"Progeny",labels=c("time","volume","Tumor Growth"))

}
