% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haus.R
\name{hausdorff}
\alias{hausdorff}
\title{Hausdorff distance}
\usage{
hausdorff(A, B)
}
\arguments{
\item{A}{A subset of \eqn{R^2}.}

\item{B}{A subset of \eqn{R^2}.}
}
\value{
Returnst the Hausdorff distance between the two sets A and B.
}
\description{
Calculates the Hausdorff distance between two subsets of \eqn{R^2}.
}
\details{
Given A and B, two non-empty subsets of a metric space \eqn{(M, d)}, their Hausdorff distance \eqn{d_H (A, B)} is:

\deqn{d_{ H}(A,B) = \max{\sup_{x \in A}\inf_{y \in B} d(x,y),\, \sup_{y \in B} \inf_{x \in A} d(x,y)}.}{d_H (A,B) = max{ sup_{x \in A} inf_{y \in B} d(x,y), sup_{y \in B} inf_{x \in A} d(x,y) } .}

Given that the curves in study are a subsets of \eqn{R^2} , \eqn{ { (t_1 , x_1 ),..., (t_n , x_n ) } }. Hence a natural choice for the metric \eqn{d} is the Euclidean distance on \eqn{R^2} .
}
