% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi_methods.R
\name{list_content.Connector_dbi}
\alias{list_content.Connector_dbi}
\title{List tables in the database}
\usage{
\method{list_content}{Connector_dbi}(connector_object, ...)
}
\arguments{
\item{connector_object}{Connector_dbi object}

\item{...}{Additional arguments passed to \link[DBI:dbListTables]{DBI::dbListTables}}
}
\value{
A \link{character} vector of table names
}
\description{
List tables in the database
}
\examples{
connector <- Connector_dbi$new(RSQLite::SQLite())
connector$write(iris, "iris")
connector$list_content()
}
