% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_other_methods.R
\name{list_content.Connector_fs}
\alias{list_content.Connector_fs}
\title{List content of the directory}
\usage{
\method{list_content}{Connector_fs}(connector_object, ...)
}
\arguments{
\item{connector_object}{Connector_fs object}

\item{...}{Arguments to pass to the list.files function}
}
\description{
List content of the directory
}
\examples{
connector <- Connector_fs$new(tempdir())
connector$write(iris, "iris.csv")
connector$list_content()
}
