% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_logger_generics.R
\name{list_content_cnt.ConnectorLogger}
\alias{list_content_cnt.ConnectorLogger}
\title{List contents Operation for ConnectorLogger class}
\usage{
\method{list_content_cnt}{ConnectorLogger}(connector_object, ...)
}
\arguments{
\item{connector_object}{The ConnectorLogger object.}

\item{...}{Additional parameters.}
}
\value{
The result of the read operation.
}
\description{
Implementation of the log_read_connector function for the ConnectorLogger
class.
}
