% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_logger_generics.R, R/cnt_logger_log_dbi.R,
%   R/cnt_logger_log_fs.R
\name{log-functions}
\alias{log-functions}
\alias{log_read_connector}
\alias{log_read_connector.default}
\alias{log_write_connector}
\alias{log_write_connector.default}
\alias{log_remove_connector}
\alias{log_remove_connector.default}
\alias{log_list_content_connector}
\alias{log_read_connector.ConnectorDBI}
\alias{log_write_connector.ConnectorDBI}
\alias{log_remove_connector.ConnectorDBI}
\alias{log_read_connector.ConnectorFS}
\alias{log_write_connector.ConnectorFS}
\alias{log_remove_connector.ConnectorFS}
\title{Connector Logging Functions}
\usage{
log_read_connector(connector_object, name, ...)

\method{log_read_connector}{default}(connector_object, name, ...)

log_write_connector(connector_object, name, ...)

\method{log_write_connector}{default}(connector_object, name, ...)

log_remove_connector(connector_object, name, ...)

\method{log_remove_connector}{default}(connector_object, name, ...)

log_list_content_connector(connector_object, ...)

\method{log_read_connector}{ConnectorDBI}(connector_object, name, ...)

\method{log_write_connector}{ConnectorDBI}(connector_object, name, ...)

\method{log_remove_connector}{ConnectorDBI}(connector_object, name, ...)

\method{log_read_connector}{ConnectorFS}(connector_object, name, ...)

\method{log_write_connector}{ConnectorFS}(connector_object, name, ...)

\method{log_remove_connector}{ConnectorFS}(connector_object, name, ...)
}
\arguments{
\item{connector_object}{The connector object to log operations for}

\item{name}{The name/identifier of the connector or resource}

\item{...}{Additional parameters passed to specific method implementations}
}
\value{
The result of the specific method implementation
}
\description{
Generic functions and methods for logging connector operations including read,
write, remove, and list operations. These functions provide a consistent
logging interface across different connector types (FS, DBI, etc.).
}
\section{Generic Functions}{

\describe{
\item{\code{log_read_connector}}{Logs reading operations from connectors}
\item{\code{log_write_connector}}{Logs writing operations to connectors}
\item{\code{log_remove_connector}}{Logs removal operations on connectors}
\item{\code{log_list_content_connector}}{Logs listing operations on connectors}
}
}

\section{Connector Types}{

The logging functions have specific implementations for:
\itemize{
\item ConnectorFS - File system connectors
\item ConnectorDBI - Database connectors
\item ConnectorLogger - Logger wrapper for any connector
}
}

\examples{
# Create a connector with logging
con <- connectors(test = connector_fs(tempdir()))
logged_con <- add_logs(con)

# Operations will be automatically logged
# log_read_connector(logged_con$test, "data.csv")
# log_write_connector(logged_con$test, "output.csv")

}
