% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_logger_add_logs.R
\name{log-utilities}
\alias{log-utilities}
\alias{add_logs}
\alias{print.ConnectorLogger}
\title{Add Logging Capability to Connections}
\usage{
add_logs(connections)

\method{print}{ConnectorLogger}(x, ...)
}
\arguments{
\item{connections}{An object of class \code{\link[=connectors]{connectors()}}. This should be a list
of connection objects to which logging capability will be added.}

\item{...}{Additional arguments passed to the next method.}
}
\value{
The modified \code{connections} object with logging capability added.
Each connection in the list will have the "ConnectorLogger" class
prepended to its existing classes.

The result of the next method in the dispatch chain.
}
\description{
This function adds logging capability to a list of connections by modifying
their class attributes. It ensures that the input is of the correct type and
registers the necessary S3 methods for logging.
}
\details{
The function performs the following steps:
\enumerate{
\item Checks if the input \code{connections} is of class "connectors".
\item Iterates through each connection in the list and prepends the "ConnectorLogger" class.
}

This method is designed to be called automatically when \code{print()} is used
on an object of class "ConnectorLogger". It uses \code{NextMethod()} to call
the next appropriate method in the method dispatch chain, allowing for
the default or any other custom print behavior to be executed.
}
\examples{
con <- connectors(
  sdtm = connector_fs(path = tempdir())
 )

logged_connections <- add_logs(con)

}
\seealso{
\code{\link{print}}
}
