% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_logger_log_dbi.R
\name{log_remove_connector.ConnectorDBI}
\alias{log_remove_connector.ConnectorDBI}
\title{Log Remove Operation for connector dbi}
\usage{
\method{log_remove_connector}{ConnectorDBI}(connector_object, name, ...)
}
\arguments{
\item{connector_object}{The ConnectorDBI object.}

\item{name}{The name of the connector.}

\item{...}{Additional parameters.}
}
\description{
Implementation of the log_remove_connector function for the
ConnectorDBI class.
}
