% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_logger_log_fs.R
\name{log_remove_connector.ConnectorFS}
\alias{log_remove_connector.ConnectorFS}
\title{Log Remove Operation for FS connector}
\usage{
\method{log_remove_connector}{ConnectorFS}(connector_object, name, ...)
}
\arguments{
\item{connector_object}{The ConnectorFS object.}

\item{name}{The name of the connector.}

\item{...}{Additional parameters.}
}
\description{
Implementation of the log_remove_connector function for the ConnectorFS
class.
}
