% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_logger_generics.R
\name{log_remove_connector}
\alias{log_remove_connector}
\title{Log Remove Connector}
\usage{
log_remove_connector(connector_object, name, ...)
}
\arguments{
\item{connector_object}{The connector object to log the removal of.}

\item{name}{The name of the connector.}

\item{...}{Additional parameters passed to the specific method implementation}
}
\value{
The result of the specific method implementation.
}
\description{
This function is a generic for logging the removal of a connector object. The
actual implementation of the logging is determined by the specific method for
the connector object's class.
}
