% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_logger_log_dbi.R
\name{log_write_connector.ConnectorDBI}
\alias{log_write_connector.ConnectorDBI}
\title{Log Write Operation for connector dbi}
\usage{
\method{log_write_connector}{ConnectorDBI}(connector_object, name, ...)
}
\arguments{
\item{connector_object}{The ConnectorDBI object.}

\item{name}{The name of the connector.}

\item{...}{Additional parameters.}
}
\description{
Implementation of the log_write_connector function for the ConnectorDBI
class.
}
