% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCoeffs.R
\name{makeCoeffs}
\alias{makeCoeffs}
\title{Make Coeffs}
\usage{
makeCoeffs(data, base = NULL, reg, dimBase, grid = NULL, pert)
}
\description{
Returns the data coefficients with respect to a base type chosen.
}
