% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectors.R
\name{nested_connectors}
\alias{nested_connectors}
\title{Create a nested connectors object}
\usage{
nested_connectors(...)
}
\arguments{
\item{...}{Any number of connectors object.}
}
\value{
A list with class "nested_connectors" containing the provided arguments.
}
\description{
This function creates a nested connectors object from the provided arguments.
}
