% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.Analisys.R
\name{PCA.Analysis}
\alias{PCA.Analysis}
\title{PCA Analysis}
\usage{
PCA.Analysis(data, dimBase = 5, save = FALSE, path = NULL)
}
\arguments{
\item{data}{CONNECTORList.}

\item{dimBase}{Dimension of the basis.}

\item{save}{If TRUE the bar plot of the PCA components is saved in a pdf file.}

\item{path}{Path to save plot to (combined with file name). If it is missing, the plot is saved in the working  directory.}
}
\value{
The bar plot of the PCA components and the vector of percentage values of the PCA components.
}
\description{
Generates a bar plot reporting the percentage values associated to each component identified by the Principal Component Analysis (PCA).
}
\examples{

GrowDataFile<-"data/1864dataset.xls"
AnnotationFile <-"data/1864info.txt"

CONNECTORList <- DataImport(GrowDataFile,AnnotationFile)

CONNECTORList<- DataTruncation(CONNECTORList,"Progeny",truncTime=60,labels = c("time","volume","Tumor Growth"))

PCA<-PCA.Analysis(CONNECTORList$Dataset)
PCA$plot

}
