% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSeparationTightness.R
\name{PlotSeparationTightness}
\alias{PlotSeparationTightness}
\title{Separation and Tightness Visualization}
\usage{
PlotSeparationTightness(clusterdata, Title = NULL, save = FALSE,
  path = NULL)
}
\arguments{
\item{Title}{Title to associate to the withinness and Betweenness plot.}

\item{save}{If TRUE the Betweenness and Withinness plot is saved in a pdf file.}

\item{path}{Path to save plot to (combined with file name). If it is missing, the plot is saved in the working  directory.}

\item{ClustCurve}{Data frame with five arguments : time, growth data, sample ID, cluster membership and feature values for each sample.}

\item{MeanCurves}{Matrix with the meancurve values on the columns according to different clusters.}
}
\value{
Returns the Withinness and Betweenness plot.
}
\description{
Visualization of the plots reporting the withinness and betweenness clustering measures.
}
\examples{

}
