% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeriesCurves.R
\name{PlotTimeSeries}
\alias{PlotTimeSeries}
\title{Time series plot}
\usage{
PlotTimeSeries(data, feature, labels = NULL, save = FALSE, path = NULL)
}
\arguments{
\item{data}{CONNECTORList. (see \code{\link{DataImport}})}

\item{feature}{The column name reported in the AnnotationFile containing the feature to be investigated.}

\item{labels}{The vector containing the text for the axis names and plot title.}

\item{save}{If TRUE then the plot is saved into a pdf file.}

\item{path}{The folder path  where the plot will be saved. If it is missing, the plot is saved in the current working  directory.}
}
\value{
PlotTimeSeries returns a list containing the line plot as ggplot object, namely PlotTimeSeries_plot, and data, i.e. the CONNECTORList.
}
\description{
Generates the line plot of the time series data. The curves are colored with respect to the feature chosen by the user and/or reported in the AnnotationFile.
}
\examples{
GrowDataFile<-"data/745dataset.xls"
AnnotationFile <-"data/745info.txt"

CONNECTORList <- DataImport(TimeSeriesFile,AnnotationFile)

PlotTimeSeries(CONNECTORList,"Progeny",labels=c("Time","Volume","Tumor Growth"))

}
\author{
Cordero Francesca, Pernice Simone, Sirovich Roberta
}
