% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotWithinness_i.R
\name{PlotWithinness.i}
\alias{PlotWithinness.i}
\title{i-th Cluster Withinness Plot}
\usage{
PlotWithinness.i(ClustCurve, MeanCurves, i, shift = 0)
}
\arguments{
\item{ClustCurve}{Data frame with five arguments : time, growth values, ID, cluster membership and feature values for each sample.}

\item{MeanCurves}{Matrix with the meancurve values on the columns according to different clusters.}

\item{i}{Value for the cluster involved in withinness computation.}

\item{shift}{Vector representing the distances between the first withinness circle center and the other centers.}

\item{centroids}{If TRUE the WithCluster_MeanDist function is used, otherwise withinness is calculated using WithCluster_CurvDist.}
}
\value{
Returns a plot representing the i-th cluster withinness measures circles.
}
\description{
Plots the i-th cluster withinness measure circles.
}
