% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotWithinnessBetweenness.R
\name{PlotWithinnessBetweenness}
\alias{PlotWithinnessBetweenness}
\title{Withinness and Betweenness Visualization}
\usage{
PlotWithinnessBetweenness(ClustCurve, MeanCurves, Title = NULL, save = TRUE,
  path = NULL)
}
\arguments{
\item{ClustCurve}{Data frame with five arguments : time, growth data, sample ID, cluster membership and feature values for each sample.}

\item{MeanCurves}{Matrix with the meancurve values on the columns according to different clusters.}

\item{Title}{Title to associate to the withinness and Betweenness plot.}

\item{save}{If TRUE the Betweenness and Withinness plot is saved in a pdf file.}

\item{path}{Path to save plot to (combined with file name). If it is missing, the plot is saved in the working  directory.}
}
\value{
Returns the Withinness and Betweenness plot.
}
\description{
Visualization of the plots reporting the withinness and betweenness clustering measures.
}
\examples{

### Data files
GrowDataFile<-"data/1864dataset.xls"
AnnotationFile <-"data/1864info.txt"

### Merge curves and target file
CONNECTORList <- DataImport(GrowDataFile,AnnotationFile)
CONNECTORList<- DataTruncation(CONNECTORList,feature="Progeny",truncTime=60,save=TRUE,path="~/Desktop/ImagesPerFrancesca/",labels = c("time","volume","Tumor Growth"

CONNECTORList.FCM <- ClusterChoice(CONNECTORList,K=c(2:6),h=2)

CONNECTORList.models <- FittingAndClustering( data= CONNECTORList, FCM_all = CONNECTORList.FCM, h = 2, k=4, feature = "Progeny", labels = c("time","volume"))

### Withinness and betweenness plot for Matlhus model and FCM

Malthus.ClustCurve <-  CONNECTORList.models$Malthus$Information$ClustCurve
Malthus.MeanCurves <-  CONNECTORList.models$Malthus$Information$meancurves
PlotWithinnessBetweenness(Malthus.ClustCurve,Malthus.MeanCurves,Title = "Malthus Cluster betweenness and withinness")

FCM.ClustCurve <-  CONNECTORList.models$FCM$Information$ClustCurve
FCM.MeanCurves <-  CONNECTORList.models$FCM$Information$meancurves

PlotWithinnessBetweenness(FCM.ClustCurve,FCM.MeanCurves,Title = "FCM Cluster betweenness and withinness")

}
