% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conts_datasources.R
\name{process_ellipsis_params}
\alias{process_ellipsis_params}
\title{Process Parameters for Functions with Ellipsis}
\usage{
process_ellipsis_params(params)
}
\arguments{
\item{params}{The extracted parameters from the function call.}
}
\value{
A list of processed parameters.
}
\description{
Handles parameter processing for functions that use ... in their arguments.
}
\keyword{internal}
