% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_logger_generics.R
\name{read_cnt.ConnectorLogger}
\alias{read_cnt.ConnectorLogger}
\title{Log Read Operation for ConnectorLogger class}
\usage{
\method{read_cnt}{ConnectorLogger}(connector_object, name, ...)
}
\arguments{
\item{connector_object}{The ConnectorLogger object.}

\item{name}{The name of the connector.}

\item{...}{Additional parameters.}
}
\value{
The result of the read operation.
}
\description{
Implementation of the log_read_connector function for the ConnectorLogger
class.
}
