% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi_methods.R
\name{remove.Connector_dbi}
\alias{remove.Connector_dbi}
\title{Remove a table from the database}
\usage{
\method{remove}{Connector_dbi}(connector_object, name, ...)
}
\arguments{
\item{connector_object}{Connector_dbi object}

\item{name}{Table name}

\item{...}{Additional arguments passed to \link[DBI:dbRemoveTable]{DBI::dbRemoveTable}}
}
\description{
Remove a table from the database
}
\examples{
connector <- Connector_dbi$new(RSQLite::SQLite())
connector$write(iris, "iris")
connector$remove("iris")
}
