% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_other_methods.R
\name{remove.Connector_fs}
\alias{remove.Connector_fs}
\title{Remove a file or directory}
\usage{
\method{remove}{Connector_fs}(connector_object, ...)
}
\arguments{
\item{connector_object}{Connector_fs object}

\item{...}{file name in case of Connector_fs}
}
\description{
Remove a file or directory
}
\examples{
connector <- Connector_fs$new(tempdir())
connector$write(iris, "iris.csv")
connector$remove("iris.csv")
}
