% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_logger_generics.R
\name{remove_cnt.ConnectorLogger}
\alias{remove_cnt.ConnectorLogger}
\title{Log Remove Operation for ConnectorLogger class}
\usage{
\method{remove_cnt}{ConnectorLogger}(connector_object, name, ...)
}
\arguments{
\item{connector_object}{The ConnectorLogger object.}

\item{name}{The name of the connector.}

\item{...}{Additional parameters.}
}
\value{
The result of the remove operation.
}
\description{
Implementation of the log_remove_connector function for the ConnectorLogger
class.
}
