% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportGeneration.R
\name{ReportGeneration}
\alias{ReportGeneration}
\title{Report generation}
\usage{
ReportGeneration(
  data,
  clusterdata,
  G,
  feature = "ID",
  file = "Report.pdf",
  truncTime = NULL,
  p = NULL,
  pRange = NULL
)
}
\arguments{
\item{data}{CONNECTORList. (see \code{\link{DataImport}})}

\item{clusterdata}{The list obtained from extrapolating the most probable clustering from the StabilityAnalysis function output. (see \code{\link{StabilityAnalysis}} and \code{\link{MostProbableClustering.Extrapolation}}).}

\item{feature}{The column name reported in the AnnotationFile  containing the feature to be investigated.}

\item{file}{}

\item{p}{The vector of the dimension of the natural cubic spline basis, by which the cross-validated loglikelihood is calculated and plotted. It could be the path of the Rds file storing the output of the \code{\link{BasisDimension.Choice}} function. If NULL, the p selection step is omitted.}

\item{path}{The folder path  where the plot will be saved. If it is missing, the plot is saved in the current working  directory.}
}
\value{

}
\description{
Report generation
}
\examples{

}
\seealso{
\code{\link{GrowthCurve}}, code{\link{TimeGridDensity}}.
}
\author{
Cordero Francesca, Pernice Simone, Sirovich Roberta
}
