% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportImages.R
\name{reportPlot.generation}
\alias{reportPlot.generation}
\title{Report plots}
\usage{
reportPlot.generation(data, crossLL = NULL, clusterData, g, title = NULL)
}
\arguments{
\item{data}{data. (see \code{\link{DataImport}})}

\item{crossLL}{the output of the cross-validated loglikelihood analysis for choosing p. (see \code{\link{BasisDimension.Choice}})}

\item{g}{Number of clusters.}

\item{title}{title of the report.}

\item{clusterdata}{The list obtained from extrapolating the most probable clustering from the StabilityAnalysis function output. (see \code{\link{StabilityAnalysis}} and \code{\link{clusterdata.Extrapolation}}).}
}
\value{

}
\description{
Generation of the report of images summarising th CONNECTOR analysis.
}
\examples{

}
\author{
Cordero Francesca, Pernice Simone, Sirovich Roberta
}
