% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunShiny.R
\name{RunConnectorShiny}
\alias{RunConnectorShiny}
\title{Shiny app}
\usage{
RunConnectorShiny()
}
\description{
function to lunch the shiny application.
}
\author{
Cordero Francesca, Pernice Simone, Sirovich Roberta
}
