% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeGridDensity.R
\name{TimeGridDensity}
\alias{TimeGridDensity}
\title{Time Grid Density}
\usage{
TimeGridDensity(data, save = FALSE, path = NULL)
}
\arguments{
\item{data}{CONNECTORList. (see \code{\link{DataImport}})}

\item{save}{If TRUE then the density time grid plot is saved into a pdf file.}

\item{path}{The folder path where the plot will be saved. If it is missing, the plot is saved in the current working  directory.}
}
\value{
Time Grid Density returns the time grid density plot as a ggplot object.
}
\description{
Plots the grid of the time points at which the growth data are collected. Each point of the grid is colored with respect to its frequency in the input data.
}
\examples{
GrowDataFile<-"data/745dataset.xls"
AnnotationFile <-"data/745info.txt"

CONNECTORList <- DataImport(GrowDataFile,AnnotationFile)
TimeGridDensity(CONNECTORList)

}
\author{
Cordero Francesca, Pernice Simone, Sirovich Roberta
}
