% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conts_datasources.R
\name{transform_as_backend}
\alias{transform_as_backend}
\title{Transform Clean Function Info to Backend Format}
\usage{
transform_as_backend(infos, name)
}
\arguments{
\item{infos}{A list with class "clean_fct_info", typically the output of \code{extract_function_info}.}

\item{name}{A character string representing the name to be assigned to the backend.}
}
\value{
A list representing the backend, with 'name' and 'backend' components or an error if the input is not of class "clean_fct_info".
}
\description{
This function takes the output of \code{extract_function_info} and transforms it
into a backend format suitable for further processing or API integration.
}
\keyword{internal}
