% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conts_datasources.R
\name{transform_as_datasources}
\alias{transform_as_datasources}
\title{Transform Multiple Backends to Datasources Format}
\usage{
transform_as_datasources(bks)
}
\arguments{
\item{bks}{A list of backends, each typically created by \code{transform_as_backend}.}
}
\value{
A list with a single 'datasources' element containing all input backends.
}
\description{
This function takes a list of backends (typically created by \code{transform_as_backend})
and wraps them in a 'datasources' list. This is useful for creating a structure
that represents multiple data sources or backends.
}
\keyword{internal}
