% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_backend.R
\name{try_connect}
\alias{try_connect}
\title{Create a backend}
\usage{
try_connect(connect_fct, params_from_user)
}
\arguments{
\item{connect_fct}{The connection function}

\item{params_from_user}{The parameters from the user}
}
\description{
Create a backend
}
