% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_logger_generics.R
\name{upload_cnt.ConnectorLogger}
\alias{upload_cnt.ConnectorLogger}
\title{Upload Operation for ConnectorLogger class}
\usage{
\method{upload_cnt}{ConnectorLogger}(
  connector_object,
  file,
  name = basename(file),
  overwrite = zephyr::get_option("overwrite", "connector"),
  ...
)
}
\arguments{
\item{connector_object}{The ConnectorLogger object.}

\item{file}{The file to upload.}

\item{name}{The name of the file in the connector.}

\item{overwrite}{Whether to overwrite existing files.}

\item{...}{Additional parameters.}
}
\value{
Invisible result of the upload operation.
}
\description{
Implementation of the upload_cnt function for the ConnectorLogger
class.
}
