% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WithCluster_CurvDist.R
\name{WithCluster_CurvDist}
\alias{WithCluster_CurvDist}
\title{Curve-Curve withinness}
\usage{
WithCluster_CurvDist(ClustCurve, i)
}
\arguments{
\item{ClustCurve}{Data frame with five arguments : time, growth values, ID, cluster membership and feature values for each sample.}

\item{i}{Value included from 0 to K (number of clusters) representing the cluster involved in withinness computation.}
}
\value{
Returns a vector reporting the hausdorff distance among the  curves belonging to the i-th cluster.
}
\description{
Computes the Hausdorff distance among curves belonging to the i-th cluster.
}
