% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WithCluster_MeanDist.R
\name{WithCluster_MeanDist}
\alias{WithCluster_MeanDist}
\title{Curve-Centroid withinness}
\usage{
WithCluster_MeanDist(ClustCurve, MeanCurves, i)
}
\arguments{
\item{ClustCurve}{Data frame with five arguments : time, growth values, ID, cluster membership and feature values for each sampple.}

\item{MeanCurves}{Matrix with the meancurves on the columns according to different clusters.}

\item{i}{Value included from 0 to K (number of clusters) representing the cluster involved in withinness computation.}
}
\value{
Returns a vector reporting the hausdorff distance of the i-th cluster curves from the corresponding cluster meancurve.
}
\description{
Computes the Hausdorff distance between the curves belonging to the i-th cluster and the corresponding mean curve.
}
