% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Withinness.R
\name{Withinness}
\alias{Withinness}
\title{Overall Clusters withinness}
\usage{
Withinness(ClustCurve, MeanCurves, centroids = TRUE)
}
\arguments{
\item{ClustCurve}{A data frame with five arguments: time, growth values, ID, cluster membership and feature values for each sample.}

\item{MeanCurves}{Matrix with the meancurve values on the columns according to different clusters.}

\item{centroids}{If TRUE the WithCluster_MeanDist function is used, otherwise withinness is calculated using WithCluster_CurvDist.}
}
\value{
Returns a matrix with four columns: mean, standard deviation, minimum and maximum withinness distance across the clusters.
}
\description{
Compute the withinness across all clusters obtained.
}
