% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi_methods.R
\name{write.Connector_dbi}
\alias{write.Connector_dbi}
\title{Write a table to the database}
\usage{
\method{write}{Connector_dbi}(connector_object, x, name, ...)
}
\arguments{
\item{connector_object}{Connector_dbi object}

\item{x}{Table to write}

\item{name}{Table name}

\item{...}{Additional arguments passed to \link[DBI:dbWriteTable]{DBI::dbWriteTable}}
}
\description{
Write a table to the database
}
\examples{
connector <- Connector_dbi$new(RSQLite::SQLite())
connector$write(iris, "iris")
}
