% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_logger_generics.R
\name{write_cnt.ConnectorLogger}
\alias{write_cnt.ConnectorLogger}
\title{Log Write Operation for ConnectorLogger class}
\usage{
\method{write_cnt}{ConnectorLogger}(connector_object, x, name, ...)
}
\arguments{
\item{connector_object}{The ConnectorLogger object.}

\item{x}{The data to write.}

\item{name}{The name of the connector.}

\item{...}{Additional parameters.}
}
\value{
Invisible result of the write operation.
}
\description{
Implementation of the log_write_connector function for the ConnectorLogger
class.
}
