% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_write.R, R/fs_write_ext_methods.R
\name{write_file}
\alias{write_file}
\alias{write_ext}
\title{Write files based on the extension}
\usage{
write_file(x, file, overwrite = FALSE, ...)
}
\arguments{
\item{x}{Object to write}

\item{file}{\code{\link[=character]{character()}} Path to write the file.}

\item{overwrite}{\link{logical} Overwrite existing content if it exists.}

\item{...}{Other parameters passed on the functions behind the methods for each file extension.}

\item{delim}{\code{\link[=character]{character()}} Delimiter to use. Default is \code{","}.}
}
\value{
\code{write_file()}: \code{\link[=invisible]{invisible()}} file.
}
\description{
\code{write_file()} is the backbone of all \code{\link[=write_cnt]{write_cnt()}} methods, where files are written
to a connector. The function is a wrapper around \code{write_ext()} where the appropriate
function to write the file is chosen depending on the file extension.

\itemize{
\item \code{txt}: \code{\link[readr:read_lines]{readr::write_lines()}}
}

\itemize{
\item \code{csv}: \code{\link[readr:write_delim]{readr::write_csv()}}
}

\itemize{
\item \code{tsv}: \code{\link[readr:write_delim]{readr::write_tsv()}}
}

\itemize{
\item \code{parquet}: \code{\link[arrow:write_parquet]{arrow::write_parquet()}}
}

\itemize{
\item \code{rds} (data.frame): \code{\link[readr:read_rds]{readr::write_rds()}}
}

\itemize{
\item \code{rds} (list): \code{\link[readr:read_rds]{readr::write_rds()}}
}

\itemize{
\item \code{rds} (character): \code{\link[readr:read_rds]{readr::write_rds()}}
}

\itemize{
\item \code{rds} (numeric): \code{\link[readr:read_rds]{readr::write_rds()}}
}

\itemize{
\item \code{rds} (logical): \code{\link[readr:read_rds]{readr::write_rds()}}
}

\itemize{
\item \code{rds} (matrix): \code{\link[readr:read_rds]{readr::write_rds()}}
}

\itemize{
\item \code{rds} (array): \code{\link[readr:read_rds]{readr::write_rds()}}
}

\itemize{
\item \code{xpt}: \code{\link[haven:read_xpt]{haven::write_xpt()}}
}

\itemize{
\item \code{yaml} (data.frame): \code{\link[yaml:write_yaml]{yaml::write_yaml()}}
}

\itemize{
\item \code{yaml} (list): \code{\link[yaml:write_yaml]{yaml::write_yaml()}}
}

\itemize{
\item \code{yaml} (character): \code{\link[yaml:write_yaml]{yaml::write_yaml()}}
}

\itemize{
\item \code{yml} (data.frame): \code{\link[yaml:write_yaml]{yaml::write_yaml()}}
}

\itemize{
\item \code{yml} (list): \code{\link[yaml:write_yaml]{yaml::write_yaml()}}
}

\itemize{
\item \code{yml} (character): \code{\link[yaml:write_yaml]{yaml::write_yaml()}}
}

\itemize{
\item \code{json} (data.frame): \code{\link[jsonlite:read_json]{jsonlite::write_json()}}
}

\itemize{
\item \code{json} (list): \code{\link[jsonlite:read_json]{jsonlite::write_json()}}
}

\itemize{
\item \code{json} (character): \code{\link[jsonlite:read_json]{jsonlite::write_json()}}
}

\itemize{
\item \code{excel}: \code{\link[writexl:write_xlsx]{writexl::write_xlsx()}}
}
}
\details{
Note that \code{write_file()} will not overwrite existing files unless \code{overwrite = TRUE},
while all methods for \code{write_ext()} will overwrite existing files by default.
}
\examples{
# Write CSV file
temp_csv <- tempfile("iris", fileext = ".csv")
write_file(iris, temp_csv)

}
