% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumberKnotsChoice.R
\name{BasisDimension.Choice}
\alias{BasisDimension.Choice}
\title{Choice of the B-spline dimension}
\usage{
BasisDimension.Choice(data, p, save = FALSE, path = NULL)
}
\arguments{
\item{data}{CONNECTORList. (see \code{\link{DataImport}})}

\item{p}{The vector of the dimension of the natural cubic spline basis.}

\item{save}{If TRUE then the plot is saved into a pdf file.}

\item{path}{The folder path  where the plot will be saved. If it is missing, the plot is saved in the current working  directory.}
}
\value{
DimensionBasis.Choice returns line plot of the cross-validated loglikelihood for each value of p, in grey the result of all ten repetitions of the likelihood calculation and in black the mean of them.
}
\description{
Generates a line plot reporting the cross-validated loglikelihood value for each number of knots. In details, for each number of knots 10\% of the curves from the whole data are removed and treated as a test set, then the remaing curves are fitted using the FCM and the loglikelihood on the test set is calculated. The process is then repeated nine more times.
}
\examples{

GrowDataFile<-"data/1864dataset.xls"
AnnotationFile <-"data/1864info.txt"

CONNECTORList <- DataImport(GrowDataFile,AnnotationFile)

CONNECTORList<- DataTruncation(CONNECTORList,"Progeny",truncTime=60,labels = c("time","volume","Tumor Growth"))

CrossLogLike<-BasisDimension.Choice(CONNECTORList,2:10)
CrossLogLike$CrossLogLikePlot



}
\references{
Gareth M. James and Catherine A. Sugar, (2000). Principal component models for sparse functional data.
}
