% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classification.R
\name{ClassificationNewCurves}
\alias{ClassificationNewCurves}
\title{Classification New Curves}
\usage{
ClassificationNewCurves(
  newdata,
  clusterdata,
  entropyCutoff = 1,
  probCutoff = 0.6,
  Cores = 1,
  Cut = F
)
}
\arguments{
\item{newdata}{Dataframe of three columns storing the new longitudinal data to classify into the clusters stored in \emph{clusterdata}. The first columns, called "\emph{ID}", stores the identification number of each sample. The second column, labeled "\emph{Observation}", contains  the data observation over the time and the respective time point is reported in the third column, labeled  "\emph{Time}"}

\item{clusterdata}{The list obtained from extrapolating the most probable clustering from the StabilityAnalysis function output. (see \code{\link{StabilityAnalysis}} and \code{\link{clusterdata.Extrapolation}}).}

\item{entropyCutoff}{Entropy filter value for the definition of the UNCLASSIFIED cluster. The default is 1, and it works together with  probCutoff parameter.}

\item{probCutoff}{Probability filter value for the definition of the UNCLASSIFIED cluster. The default is 0.6, and it works together with  entropyCutoff parameter.}

\item{Cores}{Number of cores to parallelize computations.}
}
\description{
Filter formula for the definion of the UNCLASSIFIED cluster
$ Entropy >= entropyCutoff && MajorProb<=probCutoff$ then $"Unclassified"$
}
\author{
Cordero Francesca, Pernice Simone, Sirovich Roberta
}
