% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classification.R
\name{ClassificationNewCurves}
\alias{ClassificationNewCurves}
\title{Classification New Curves}
\usage{
ClassificationNewCurves(newdata, clusterdata)
}
\arguments{
\item{newdata}{}

\item{clusterdata}{The list obtained from extrapolating the most probable clustering from the StabilityAnalysis function output. (see \code{\link{StabilityAnalysis}} and \code{\link{clusterdata.Extrapolation}}).}
}
\value{

}
\description{

}
\examples{

}
\author{
Cordero Francesca, Pernice Simone, Sirovich Roberta
}
