% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterWithMeanCurves.R
\name{ClusterWithMeanCurve}
\alias{ClusterWithMeanCurve}
\title{Mean Cluster Curves}
\usage{
ClusterWithMeanCurve(clusterdata, data, feature, title = "", labels = c("",
  ""), path = NULL, save = FALSE)
}
\arguments{
\item{data}{CONNECTORList.}

\item{feature}{The column name reported in the AnnotationFile containing the feature interesting for the user to be investigated.}

\item{labels}{Vector containing the text for the title of axis.}

\item{clusterdata.FCM}{Object belonging to the class funcyOutList.}

\item{clusterdata.Models}{Object, or a list of objects storing the parameters, clusters, centers, meancurves and seed derived from fitting and clustering the data using Malthus, Gompertz or/and Logistic model.}

\item{model}{Model name, i.e. FCM, Malthus, Gompertz or Logistic. The default is "ALL" that runs all the models but it needs in input both clusterdata.FCM and clusterdata.Models .}
}
\value{
The plot reporting the mean curves of all clusters, for each cluster the plot of the growth curves and a list of data reporting the mean curve values and for each sample the ID of the belonging cluster.
}
\description{
Visualize the plot of the mean cluster curves given one model (i.e. FCM, Malthus, Gompertz or Logistic).
ClusterWithMeanCurve  generates k plots, where k is the number of clusters: in each plot the curves are colored according to the feature chosen by the user.
}
\examples{

GrowDataFile<-"data/1864dataset.xls"
AnnotationFile <-"data/1864info.txt"

CONNECTORList <- DataImport(GrowDataFile,AnnotationFile).............

}
