% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_generics.R
\name{cnt_create_directory}
\alias{cnt_create_directory}
\title{Create a directory}
\usage{
cnt_create_directory(connector_object, name, ...)
}
\arguments{
\item{name}{\link{character} The name of the directory to create}

\item{...}{Additional arguments passed to the method for the individual connector.}
}
\description{
Create a directory
}
