% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_generics.R, R/fs_methods.R
\name{cnt_download}
\alias{cnt_download}
\alias{cnt_download.connector_fs}
\title{Download content from the connector}
\usage{
cnt_download(connector_object, name, file = basename(name), ...)

\method{cnt_download}{connector_fs}(connector_object, name, file = basename(name), ...)
}
\arguments{
\item{connector_object}{\link{connector} The connector object to use.}

\item{name}{\link{character} Name of the content to read, write, or remove. Typically the table name.}

\item{file}{\link{character} Path to the file to download to or upload from}

\item{...}{Additional arguments passed to the method for the individual connector.}
}
\value{
\link{invisible} file.
}
\description{
Generic implementing how to download files from a connector:

\itemize{
\item \link{connector_fs}: Uses \link{file.copy} to copy a file from the file storage to the desired \code{file}.
}
}
\examples{
# Download file from a file storage
cnt_fs <- connector_fs$new(tempdir())

cnt_fs |>
  cnt_write("this is an example", "example.txt")

list.files(pattern = "example.txt")

cnt_fs |>
  cnt_download("example.txt")

list.files(pattern = "example.txt")
readLines("example.txt")

}
