% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi_methods.R
\name{cnt_list_content.connector_dbi}
\alias{cnt_list_content.connector_dbi}
\title{List tables in the database}
\usage{
\method{cnt_list_content}{connector_dbi}(connector_object, ...)
}
\arguments{
\item{connector_object}{connector_dbi object}

\item{...}{Additional arguments passed to \link[DBI:dbListTables]{DBI::dbListTables}}
}
\value{
A \link{character} vector of table names
}
\description{
List tables in the database
}
\examples{
connector <- connector_dbi$new(RSQLite::SQLite())
connector$cnt_write(iris, "iris")
connector$cnt_list_content()
}
