% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_methods.R
\name{cnt_list_content.connector_fs}
\alias{cnt_list_content.connector_fs}
\title{List content of the directory}
\usage{
\method{cnt_list_content}{connector_fs}(connector_object, ...)
}
\arguments{
\item{connector_object}{connector_fs object}

\item{...}{Arguments to pass to the list.files function}
}
\description{
List content of the directory
}
\examples{
connector <- connector_fs$new(tempdir())
connector$cnt_write(iris, "iris.csv")
connector$cnt_list_content()
}
