% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_generics.R, R/fs_methods.R
\name{cnt_list_content}
\alias{cnt_list_content}
\alias{cnt_list_content.connector_fs}
\title{List available content from the connector}
\usage{
cnt_list_content(connector_object, ...)

\method{cnt_list_content}{connector_fs}(connector_object, ...)
}
\arguments{
\item{connector_object}{\link{connector} The connector object to use.}

\item{...}{Additional arguments passed to the method for the individual connector.}
}
\value{
A \link{character} vector of content names
}
\description{
Generic implementing how to list all content available for different connectors:

\itemize{
\item \link{connector_fs}: Uses \link{list.files} to list all files at the path of the connector.
}
}
\examples{
# List content in a file storage
cnt_fs <- connector_fs$new(tempdir())

cnt_fs |>
  cnt_list_content()

# Only list CSV files using the pattern argument of list.files

cnt_fs |>
  cnt_list_content(pattern = "\\\\.csv$")

}
