% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_generics.R, R/fs_methods.R
\name{cnt_list_content}
\alias{cnt_list_content}
\alias{cnt_list_content.connector_fs}
\title{List available content from the connector}
\usage{
cnt_list_content(connector_object, ...)

\method{cnt_list_content}{connector_fs}(connector_object, ...)
}
\arguments{
\item{connector_object}{connector_fs object}

\item{...}{Arguments to pass to the list.files function}
}
\value{
A \link{character} vector of content names
}
\description{
List available content from the connector

List content of the directory
}
\examples{
connector <- connector_fs$new(tempdir())
connector$cnt_write(iris, "iris.csv")
connector$cnt_list_content()
}
