% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_generics.R
\name{cnt_read}
\alias{cnt_read}
\title{Read content from the connector}
\usage{
cnt_read(connector_object, ...)
}
\arguments{
\item{connector_object}{\link{connector} The connector object to use.}

\item{...}{Additional arguments passed to the method for the individual connector.}
}
\value{
R object with the content. For rectangular data a \link{data.frame}.
}
\description{
Read content from the connector
}
