% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_generics.R, R/fs_methods.R
\name{cnt_remove}
\alias{cnt_remove}
\alias{cnt_remove.connector_fs}
\title{Remove content from the connector}
\usage{
cnt_remove(connector_object, name, ...)

\method{cnt_remove}{connector_fs}(connector_object, name, ...)
}
\arguments{
\item{connector_object}{\link{connector} The connector object to use.}

\item{name}{\link{character} Name of the content to read, write, or remove. Typically the table name.}

\item{...}{Additional arguments passed to the method for the individual connector.}
}
\value{
\link{invisible} connector_object.
}
\description{
Generic implementing how to remove content from different connectors:

\itemize{
\item \link{connector_fs}: Uses \link{unlink} to delete the file.
}
}
\examples{
# Remove a file from the file storage
cnt_fs <- connector_fs$new(tempdir())

cnt_fs |>
  cnt_write("this is an example", "example.txt")

cnt_fs |>
  cnt_remove("example.txt")

cnt_fs |>
  cnt_remove("example.txt")

cnt_fs |>
  cnt_list_content(pattern = "example.txt")

}
