% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_generics.R
\name{cnt_remove_directory}
\alias{cnt_remove_directory}
\title{Remove a directory}
\usage{
cnt_remove_directory(connector_object, name, ...)
}
\arguments{
\item{name}{\link{character} The name of the directory to remove}

\item{...}{Additional arguments passed to the method for the individual connector.}
}
\description{
Remove a directory
}
