% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi_methods.R
\name{cnt_tbl.connector_dbi}
\alias{cnt_tbl.connector_dbi}
\title{Create a \link{tbl} object}
\usage{
\method{cnt_tbl}{connector_dbi}(connector_object, name, ...)
}
\arguments{
\item{connector_object}{connector_dbi object}

\item{name}{Table name}

\item{...}{Additional arguments passed to \link[dplyr:tbl]{dplyr::tbl}}
}
\description{
Create a \link{tbl} object
}
\examples{
connector <- connector_dbi$new(RSQLite::SQLite())
connector$cnt_write(iris, "iris")
connector$cnt_tbl("iris")
}
