% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_generics.R
\name{cnt_upload}
\alias{cnt_upload}
\title{Upload content to the connector}
\usage{
cnt_upload(connector_object, file, name = basename(file), ...)
}
\arguments{
\item{connector_object}{\link{connector} The connector object to use.}

\item{file}{\link{character} Path to the file to download to or upload from}

\item{name}{\link{character} Name of the content to read, write, or remove. Typically the table name.}

\item{...}{Additional arguments passed to the method for the individual connector.}
}
\value{
\link{invisible} \code{connector_object}
}
\description{
Upload content to the connector
}
