% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_generics.R
\name{cnt_write}
\alias{cnt_write}
\title{Write content to the connector}
\usage{
cnt_write(connector_object, x, ...)
}
\arguments{
\item{connector_object}{\link{connector} The connector object to use.}

\item{x}{The object to write to the connection}

\item{...}{Additional arguments passed to the method for the individual connector.}
}
\value{
\link{invisible} \code{connector_object}
}
\description{
Write content to the connector
}
