% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectors.R
\name{connect}
\alias{connect}
\title{Connect to datasources specified in config file}
\usage{
connect(config = "_connector.yml")
}
\arguments{
\item{config}{\link{character} path to a connector config file or a \link{list} of specifications}
}
\value{
\link{connectors}
}
\description{
Connect to datasources specified in config file
}
\examples{
config <- system.file("config", "default_config.yml", package = "connector")
con <- connect(config)
con
}
