% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect}
\alias{connect}
\title{Connect to datasources specified in a config file}
\usage{
connect(config = "_connector.yml", set_env = TRUE)
}
\arguments{
\item{config}{\link{character} path to a connector config file or a \link{list} of specifications}

\item{set_env}{\link{logical} Should environment variables from the yaml file be set. Default is TRUE.}
}
\value{
\link{connectors}
}
\description{
Based on a configuration file or list this functions creates a \code{\link[=connectors]{connectors()}} object with
a \link{connector} for each of the specified datasources.

The configuration file can be in any format that can be read through \code{\link[=read_file]{read_file()}}, and
contains a list.

See also \code{vignette("connector")} on how to use configuration files in your project,
details below for the required structure of the configuration.
}
\details{
The input list have to have the following structure:
\itemize{
\item Only metadata, env, connections, and datasources are allowed.
\item All elements must be named.
\item \strong{connections} and \strong{datasources} are mandatory.
\item \strong{metadata} and \strong{env} must each be a list of named character vectors of length 1 if specified.
\item \strong{connections} and \strong{datasources} must each be a list of unnamed lists.
\item Each connection must have the named character element \strong{con} and the named list element \strong{backend}
\item Each datasource must have the named character element \strong{name}
\item For each connection \strong{backend}.\strong{type} must be provided
}
}
\examples{
config <- system.file("config", "default_config.yml", package = "connector")

# Show the raw configuration file
readLines(config) |>
 cat(sep = "\n")

# Connect to the datasources specified in it
cnts <- connect(config)
cnts

# Content of each connector

cnts$adam
cnts$sdtm

}
