% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.R
\name{connect_dbi}
\alias{connect_dbi}
\title{Create DBI connector}
\usage{
connect_dbi(drv, ...)
}
\arguments{
\item{drv}{DBI driver}

\item{...}{Additional arguments passed to \link[DBI:dbConnect]{DBI::dbConnect}}
}
\value{
A new \link{connector_dbi} object
}
\description{
Create DBI connector
}
