% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector-options.R
\name{connector-options-params}
\alias{connector-options-params}
\title{Internal parameters for reuse in functions}
\arguments{
\item{verbosity_level}{Verbosity level for functions in connector.
See \link[zephyr:verbosity_level]{zephyr::verbosity_level} for details.. Default: \code{"verbose"}.}

\item{overwrite}{Overwrite existing content if it exists in the connector?
See \link{connector-options} for details. Default can be set globally with
\code{options(connector.overwrite = TRUE/FALSE)} or environment variable
\code{R_CONNECTOR_OVERWRITE}.. Default: \code{FALSE}.}

\item{logging}{Add logs to the console as well as to the whirl log html files. Default: \code{FALSE}.}

\item{default_ext}{Default extension to use when writing and reading files. Default: \code{"rds"}.}
}
\description{
Internal parameters for reuse in functions
}
\details{
See \link{connector-options} for more information.
}
\keyword{internal}
