% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector-options.R
\name{connector-options}
\alias{connector-options}
\title{Options for connector}
\description{
\subsection{verbosity_level}{

Verbosity level for functions in connector.
See \link[zephyr:verbosity_level]{zephyr::verbosity_level} for details.
\itemize{
\item Default: \code{"verbose"}
\item Option: \code{connector.verbosity_level}
\item Environment: \code{R_CONNECTOR_VERBOSITY_LEVEL}
}
}

\subsection{overwrite}{

Overwrite existing content if it exists in the connector?
\itemize{
\item Default: \code{FALSE}
\item Option: \code{connector.overwrite}
\item Environment: \code{R_CONNECTOR_OVERWRITE}
}
}

\subsection{logging}{

Add logs to the console as well as to the whirl log html files
\itemize{
\item Default: \code{FALSE}
\item Option: \code{connector.logging}
\item Environment: \code{R_CONNECTOR_LOGGING}
}
}
}
