% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector-package.R
\docType{package}
\name{connector-package}
\alias{connector-package}
\title{connector: Connect to your data easily}
\description{
The connector package helps connecting R to databases that contains ADAM, STDM and so on. The package defines an interface that can be reuse for any databases.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://novonordisk-opensource.github.io/connector}
  \item \url{https://github.com/novonordisk-opensource/connector}
}

}
\author{
\strong{Maintainer}: Cervan Girard \email{cgid@novonordisk.com}

Authors:
\itemize{
  \item Aksel Thomsen \email{oath@novonordisk.com}
}

Other contributors:
\itemize{
  \item Novo Nordisk A/S [copyright holder]
}

}
\keyword{internal}
