% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector.R
\name{connector}
\alias{connector}
\title{General connector object}
\description{
This R6 class is a general class for all connectors. It is used to define the methods that all connectors should have.
New connectors should inherit from this class, and the methods described in \link{connector_methods} should be implemented.
}
\seealso{
\link{connector_methods} \link{connector_fs} and \link{connector_dbi} for examples on how to create a new connector class.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-connector-list_content}{\code{connector$list_content()}}
\item \href{#method-connector-read}{\code{connector$read()}}
\item \href{#method-connector-write}{\code{connector$write()}}
\item \href{#method-connector-remove}{\code{connector$remove()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connector-list_content"></a>}}
\if{latex}{\out{\hypertarget{method-connector-list_content}{}}}
\subsection{Method \code{list_content()}}{
List content
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{connector$list_content(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments passed to the method for the individual connector.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \link{character} vector of content names
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connector-read"></a>}}
\if{latex}{\out{\hypertarget{method-connector-read}{}}}
\subsection{Method \code{read()}}{
Read content
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{connector$read(name, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{\link{character} Name of the content to read, write, or remove. Typically the table name,}

\item{\code{...}}{Additional arguments passed to the method for the individual connector.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The result of the read method
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connector-write"></a>}}
\if{latex}{\out{\hypertarget{method-connector-write}{}}}
\subsection{Method \code{write()}}{
Write content
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{connector$write(x, name, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The object to write to the connection}

\item{\code{name}}{\link{character} Name of the content to read, write, or remove. Typically the table name,}

\item{\code{...}}{Additional arguments passed to the method for the individual connector.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connector-remove"></a>}}
\if{latex}{\out{\hypertarget{method-connector-remove}{}}}
\subsection{Method \code{remove()}}{
Remove or delete content
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{connector$remove(name, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{\link{character} Name of the content to read, write, or remove. Typically the table name,}

\item{\code{...}}{Additional arguments passed to the method for the individual connector.}
}
\if{html}{\out{</div>}}
}
}
}
