% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector.R
\name{connector}
\alias{connector}
\title{General connector object}
\description{
This R6 class is a general class for all connectors.
It is used to define the methods that all connectors should have.
New connectors should inherit from this class,
and the methods described in \link{connector_methods} should be implemented.
}
\seealso{
\code{vignette("customize")} on how to create custom connectors and methods,
and how concrete examples in \link{connector_methods}, \link{connector_fs} and \link{connector_dbi}.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-connector-new}{\code{connector$new()}}
\item \href{#method-connector-print}{\code{connector$print()}}
\item \href{#method-connector-cnt_list_content}{\code{connector$cnt_list_content()}}
\item \href{#method-connector-cnt_read}{\code{connector$cnt_read()}}
\item \href{#method-connector-cnt_write}{\code{connector$cnt_write()}}
\item \href{#method-connector-cnt_remove}{\code{connector$cnt_remove()}}
\item \href{#method-connector-clone}{\code{connector$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connector-new"></a>}}
\if{latex}{\out{\hypertarget{method-connector-new}{}}}
\subsection{Method \code{new()}}{
Initialize connector with the option of adding an extra class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{connector$new(extra_class = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{extra_class}}{\link{character} Extra class to assign to the new connector.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connector-print"></a>}}
\if{latex}{\out{\hypertarget{method-connector-print}{}}}
\subsection{Method \code{print()}}{
Print method for a connector showing the registered methods
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{connector$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Invisible self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connector-cnt_list_content"></a>}}
\if{latex}{\out{\hypertarget{method-connector-cnt_list_content}{}}}
\subsection{Method \code{cnt_list_content()}}{
List content
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{connector$cnt_list_content(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments passed to the method for the individual connector.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \link{character} vector of content names
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connector-cnt_read"></a>}}
\if{latex}{\out{\hypertarget{method-connector-cnt_read}{}}}
\subsection{Method \code{cnt_read()}}{
Read content
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{connector$cnt_read(name, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{\link{character} Name of the content to read, write, or remove. Typically the table name.}

\item{\code{...}}{Additional arguments passed to the method for the individual connector.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The result of the read method
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connector-cnt_write"></a>}}
\if{latex}{\out{\hypertarget{method-connector-cnt_write}{}}}
\subsection{Method \code{cnt_write()}}{
Write content
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{connector$cnt_write(x, name, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The object to write to the connection}

\item{\code{name}}{\link{character} Name of the content to read, write, or remove. Typically the table name.}

\item{\code{...}}{Additional arguments passed to the method for the individual connector.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connector-cnt_remove"></a>}}
\if{latex}{\out{\hypertarget{method-connector-cnt_remove}{}}}
\subsection{Method \code{cnt_remove()}}{
Remove or delete content
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{connector$cnt_remove(name, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{\link{character} Name of the content to read, write, or remove. Typically the table name.}

\item{\code{...}}{Additional arguments passed to the method for the individual connector.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connector-clone"></a>}}
\if{latex}{\out{\hypertarget{method-connector-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{connector$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
