% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector.R
\name{connector}
\alias{connector}
\title{R6 general connector class}
\description{
R6 general connector class

R6 general connector class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-connector-new}{\code{connector$new()}}
\item \href{#method-connector-connect}{\code{connector$connect()}}
\item \href{#method-connector-disconnect}{\code{connector$disconnect()}}
\item \href{#method-connector-clone}{\code{connector$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connector-new"></a>}}
\if{latex}{\out{\hypertarget{method-connector-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{connector$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connector-connect"></a>}}
\if{latex}{\out{\hypertarget{method-connector-connect}{}}}
\subsection{Method \code{connect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{connector$connect()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connector-disconnect"></a>}}
\if{latex}{\out{\hypertarget{method-connector-disconnect}{}}}
\subsection{Method \code{disconnect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{connector$disconnect()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connector-clone"></a>}}
\if{latex}{\out{\hypertarget{method-connector-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{connector$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
