% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.R
\name{connector_dbi}
\alias{connector_dbi}
\title{R6 class for a dbi connection, see \link{connector_dbi} (used to interact with DBI compliant database backends)}
\description{
Connector object for DBI connections. This object is used to interact with DBI compliant database backends.
See the \href{https://dbi.r-dbi.org/}{DBI package} for how which backends are supported.
}
\details{
Upon garbage collection, the connection will try to disconnect from the database.
But it is good practice to call \code{disconnect} when you are done with the connection.
}
\examples{
# Create DBI connector

db <- connector_dbi$new(RSQLite::SQLite(), ":memory:")

db

# Write to the database

db$cnt_write(iris, "iris")

# Read from the database

db$cnt_read("iris") |>
  head(5)

# List available tables

db$cnt_list_content()

# Use the connector to run a query

db$conn

db$conn |>
  DBI::dbGetQuery("SELECT * FROM iris limit 5")

# Use dplyr verbs and collect data

db$cnt_tbl("iris") |>
  dplyr::filter(Sepal.Length > 7) |>
  dplyr::collect()

# Disconnect from the database

db$cnt_disconnect()

}
\section{Super class}{
\code{\link[connector:connector]{connector::connector}} -> \code{connector_dbi}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{conn}}{The DBI connector object of the connector}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-connector_dbi-new}{\code{connector_dbi$new()}}
\item \href{#method-connector_dbi-cnt_disconnect}{\code{connector_dbi$cnt_disconnect()}}
\item \href{#method-connector_dbi-cnt_tbl}{\code{connector_dbi$cnt_tbl()}}
\item \href{#method-connector_dbi-clone}{\code{connector_dbi$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="connector" data-topic="connector" data-id="cnt_list_content"><a href='../../connector/html/connector.html#method-connector-cnt_list_content'><code>connector::connector$cnt_list_content()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connector" data-topic="connector" data-id="cnt_read"><a href='../../connector/html/connector.html#method-connector-cnt_read'><code>connector::connector$cnt_read()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connector" data-topic="connector" data-id="cnt_remove"><a href='../../connector/html/connector.html#method-connector-cnt_remove'><code>connector::connector$cnt_remove()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connector" data-topic="connector" data-id="cnt_write"><a href='../../connector/html/connector.html#method-connector-cnt_write'><code>connector::connector$cnt_write()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connector_dbi-new"></a>}}
\if{latex}{\out{\hypertarget{method-connector_dbi-new}{}}}
\subsection{Method \code{new()}}{
Initialize the connection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{connector_dbi$new(drv, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{drv}}{DBI driver}

\item{\code{...}}{Additional arguments passed to \link[DBI:dbConnect]{DBI::dbConnect}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \link{connector_dbi} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connector_dbi-cnt_disconnect"></a>}}
\if{latex}{\out{\hypertarget{method-connector_dbi-cnt_disconnect}{}}}
\subsection{Method \code{cnt_disconnect()}}{
Disconnect from the database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{connector_dbi$cnt_disconnect()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connector_dbi-cnt_tbl"></a>}}
\if{latex}{\out{\hypertarget{method-connector_dbi-cnt_tbl}{}}}
\subsection{Method \code{cnt_tbl()}}{
Create a \link{tbl} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{connector_dbi$cnt_tbl(name, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{\link{character} Table name}

\item{\code{...}}{Additional arguments passed to \link[dplyr:tbl]{dplyr::tbl}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connector_dbi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-connector_dbi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{connector_dbi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
