% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.R
\name{connector_dbi_methods}
\alias{connector_dbi_methods}
\alias{cnt_disconnect}
\alias{cnt_tbl}
\title{Additional methods DBI connectors}
\usage{
cnt_disconnect(connector_object, ...)

cnt_tbl(connector_object, ...)
}
\arguments{
\item{connector_object}{A \link{connector_dbi} object to be able to use functions from it}

\item{...}{Additional arguments passed to the methods}
}
\description{
These methods are additional S3 methods for  \link{connector_dbi}.
}
\seealso{
\link{connector_methods}
}
