% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{connector_fs}
\alias{connector_fs}
\title{Create FS connector}
\usage{
connector_fs(path, ..., extra_class = NULL)
}
\arguments{
\item{path}{Path to the file system}

\item{...}{Additional arguments passed to \link{Connector_fs}}

\item{extra_class}{\link{character} Extra class added to the object. See details.}
}
\value{
A new \link{Connector_fs} object
}
\description{
Create a new FS connector object. See \link{Connector_fs} for details.
}
\details{
The \code{extra_class} parameter allows you to create a subclass of the \code{Connector_fs} object.
This can be useful if you want to create a custom connection object for easier dispatch of new s3 methods,
while still inheriting the methods from the \code{Connector_fs} object.
}
\examples{
# Connect to a file system

path_to_adam <- system.file("demo_trial", "adam", package = "connector")
db <- connector_fs(path_to_adam)

db

# Create subclass connection

db_subclass <- connector_fs(path_to_adam, extra_class = "subclass")

db_subclass
class(db_subclass)

}
