% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{connector_fs}
\alias{connector_fs}
\title{Connector for file storage}
\description{
The connector_fs class is a file storage connector for accessing and manipulating files any file storage solution.
The default implementation includes methods for files stored on local or network drives.
}
\section{Super class}{
\code{\link[connector:connector]{connector::connector}} -> \code{connector_fs}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Path to the file storage}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-connector_fs-new}{\code{connector_fs$new()}}
\item \href{#method-connector_fs-cnt_download}{\code{connector_fs$cnt_download()}}
\item \href{#method-connector_fs-cnt_upload}{\code{connector_fs$cnt_upload()}}
\item \href{#method-connector_fs-cnt_create_directory}{\code{connector_fs$cnt_create_directory()}}
\item \href{#method-connector_fs-cnt_remove_directory}{\code{connector_fs$cnt_remove_directory()}}
\item \href{#method-connector_fs-clone}{\code{connector_fs$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="connector" data-topic="connector" data-id="cnt_list_content"><a href='../../connector/html/connector.html#method-connector-cnt_list_content'><code>connector::connector$cnt_list_content()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connector" data-topic="connector" data-id="cnt_read"><a href='../../connector/html/connector.html#method-connector-cnt_read'><code>connector::connector$cnt_read()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connector" data-topic="connector" data-id="cnt_remove"><a href='../../connector/html/connector.html#method-connector-cnt_remove'><code>connector::connector$cnt_remove()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connector" data-topic="connector" data-id="cnt_write"><a href='../../connector/html/connector.html#method-connector-cnt_write'><code>connector::connector$cnt_write()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connector" data-topic="connector" data-id="print"><a href='../../connector/html/connector.html#method-connector-print'><code>connector::connector$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connector_fs-new"></a>}}
\if{latex}{\out{\hypertarget{method-connector_fs-new}{}}}
\subsection{Method \code{new()}}{
Initializes the connector_fs class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{connector_fs$new(path, access = "rw", extra_class = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path to the file storage}

\item{\code{access}}{Access type ("rw" by default)}

\item{\code{extra_class}}{\link{character} Extra class to assign to the new connector.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connector_fs-cnt_download"></a>}}
\if{latex}{\out{\hypertarget{method-connector_fs-cnt_download}{}}}
\subsection{Method \code{cnt_download()}}{
Read content
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{connector_fs$cnt_download(name, file = basename(name), ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{\link{character} Name of the content to read, write, or remove. Typically the table name.}

\item{\code{file}}{\link{character} The file to write or save to from the file storage.}

\item{\code{...}}{Additional arguments passed to the method.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The result of the read method
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connector_fs-cnt_upload"></a>}}
\if{latex}{\out{\hypertarget{method-connector_fs-cnt_upload}{}}}
\subsection{Method \code{cnt_upload()}}{
Upload a file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{connector_fs$cnt_upload(file, name = basename(file), ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{\link{character} The file to write or save to from the file storage.}

\item{\code{name}}{\link{character} Name of the content to read, write, or remove. Typically the table name.}

\item{\code{...}}{Additional arguments passed to the method.}

\item{\code{x}}{The object to upload to the file storage}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connector_fs-cnt_create_directory"></a>}}
\if{latex}{\out{\hypertarget{method-connector_fs-cnt_create_directory}{}}}
\subsection{Method \code{cnt_create_directory()}}{
Create a directory
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{connector_fs$cnt_create_directory(name, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the directory to create}

\item{\code{...}}{Additional arguments passed to the method.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The directory created
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connector_fs-cnt_remove_directory"></a>}}
\if{latex}{\out{\hypertarget{method-connector_fs-cnt_remove_directory}{}}}
\subsection{Method \code{cnt_remove_directory()}}{
Remove a directory
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{connector_fs$cnt_remove_directory(name, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the directory to create}

\item{\code{...}}{Additional arguments passed to the method.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connector_fs-clone"></a>}}
\if{latex}{\out{\hypertarget{method-connector_fs-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{connector_fs$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
