% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_methods.R
\name{connector_methods}
\alias{connector_methods}
\alias{read}
\alias{write}
\alias{remove}
\alias{list_content}
\alias{get_conn}
\alias{disconnect}
\alias{tbl}
\title{Defaults method for connector object}
\usage{
read(connector_object, ...)

write(connector_object, ...)

remove(connector_object, ...)

list_content(connector_object, ...)

get_conn(connector_object, ...)

disconnect(connector_object, ...)

tbl(connector_object, ...)
}
\arguments{
\item{connector_object}{A connector object to be able to use functions from it}

\item{...}{Additional arguments passed to the method}
}
\value{
The result of the read method
}
\description{
Those methods are a S3 method that dispatches to the specific methods for the connector object.
}
\details{
For example, if you have a connector object \code{Connector_fs}, you can use \code{read(Connector_fs, "file.csv")} to read the file. It will be dispatch to the \code{read.Connector_fs} method.
Why ? The main aim is to allow the user to use the same function for different connector objects by using the builder function \code{connector_fs} and add an "extra_class".
By doing so, you can create a subclass of the \code{Connector_fs} object and dispatch to the specific methods for this subclass.
For example, if you have a subclass \code{subclass}, you can use \code{read(subclass, "file.csv")} to read the file. It will be dispatch to the \code{read.subclass} method. And you can still use the \code{Connector_fs} methods.
}
