% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector.R
\name{Connector_object}
\alias{Connector_object}
\alias{Connector}
\title{General Connector object}
\description{
This R6 class is a general class for all connectors. It is used to define the methods that all connectors should have.
New connectors should inherit from this class, and the methods described in \link{connector_methods} should be implemented.
}
\seealso{
\link{connector_methods} \link{Connector_fs} and \link{Connector_dbi} for examples on how to create a new connector class.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Connector-list_content}{\code{Connector$list_content()}}
\item \href{#method-Connector-read}{\code{Connector$read()}}
\item \href{#method-Connector-write}{\code{Connector$write()}}
\item \href{#method-Connector-remove}{\code{Connector$remove()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connector-list_content"></a>}}
\if{latex}{\out{\hypertarget{method-Connector-list_content}{}}}
\subsection{Method \code{list_content()}}{
List content
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connector$list_content(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments passed to the method for the individual connector.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \link{character} vector of content names
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connector-read"></a>}}
\if{latex}{\out{\hypertarget{method-Connector-read}{}}}
\subsection{Method \code{read()}}{
Read content
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connector$read(name, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{\link{character} Name of the content to read, write, or remove. Typically the table name,}

\item{\code{...}}{Additional arguments passed to the method for the individual connector.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The result of the read method
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connector-write"></a>}}
\if{latex}{\out{\hypertarget{method-Connector-write}{}}}
\subsection{Method \code{write()}}{
Write content
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connector$write(x, name, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The object to write to the connection}

\item{\code{name}}{\link{character} Name of the content to read, write, or remove. Typically the table name,}

\item{\code{...}}{Additional arguments passed to the method for the individual connector.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connector-remove"></a>}}
\if{latex}{\out{\hypertarget{method-Connector-remove}{}}}
\subsection{Method \code{remove()}}{
Remove or delete content
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connector$remove(name, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{\link{character} Name of the content to read, write, or remove. Typically the table name,}

\item{\code{...}}{Additional arguments passed to the method for the individual connector.}
}
\if{html}{\out{</div>}}
}
}
}
