% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectors.R
\name{connectors}
\alias{connectors}
\title{Collection of connector objects}
\usage{
connectors(...)
}
\arguments{
\item{...}{Named individual \link{Connector} objects}
}
\description{
Holds a special list of individual connector objects for consistent use of
connections in your project.
}
\examples{
# Create connectors objects

cnts <- connectors(
  sdtm = connector_fs(path = tempdir()),
  adam = connector_dbi(drv = RSQLite::SQLite())
)

# Print for overview

cnts

# Print the individual connector for more information

cnts$sdtm

cnts$adam

}
