% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectors.R
\name{connectors}
\alias{connectors}
\title{Special list of connectors}
\usage{
connectors(connectors)
}
\arguments{
\item{connectors}{Named \link{list} of individual \link{connector} objects}
}
\description{
Special list of connectors
}
