% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectors.R
\name{Connectors}
\alias{Connectors}
\title{Special list of Connectors}
\usage{
Connectors(connectors)
}
\arguments{
\item{connector}{Named \link{list} of individual \link{Connector} objects}
}
\description{
Special list of Connectors
}
