% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectors.R
\name{connectors}
\alias{connectors}
\title{Collection of connector objects}
\usage{
connectors(...)
}
\arguments{
\item{...}{Named individual \link{connector} objects}
}
\description{
Holds a special list of individual connector objects for consistent use of
connections in your project.
}
\examples{
# Create connectors objects

con <- connectors(
  sdtm = ConnectorFS$new(path = tempdir()),
  adam = ConnectorDBI$new(drv = RSQLite::SQLite())
)

# Print for overview

con

# Print the individual connector for more information

con$sdtm

con$adam

}
