% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_backend.R
\name{create_backend}
\alias{create_backend}
\title{Create a backend}
\usage{
create_backend(yaml_content, backend, name)
}
\arguments{
\item{yaml_content}{The yaml content}

\item{backend}{The backend to create}

\item{name}{The name of the connection}
}
\value{
A new backend based on R6 class
}
\description{
Create a backend
}
\examples{
yaml_file <- system.file("config", "example_for_generic.yml", package = "connector")
yaml_content <- yaml::read_yaml(yaml_file, eval.expr = TRUE)
only_one <- extract_connections(yaml_content)[[1]]
## Extract fct
my_backend <- only_one \%>\%
    extract_backends()
name <- only_one \%>\%
    extract_con()
# Create the backend
test <- create_backend(yaml_content = yaml_content, backend = my_backend, name = name)

}
