% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_backend_tools.R
\name{create_backend_fs}
\alias{create_backend_fs}
\title{Create a backend for the file system}
\usage{
create_backend_fs(backend)
}
\arguments{
\item{backend}{The backend to create and because of file system, a "path" is mandatory}
}
\value{
A new backend based on R6 class
}
\description{
Create a backend for the file system
}
\examples{
yaml_file <- system.file("config", "default_config.yml", package = "connector")
yaml_content <- read_yaml_config(yaml_file)

only_one <- yaml_content[["connections"]][[1]][["backend"]]

test <- create_backend_fs(only_one)

}
