% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_generics.R, R/fs_methods.R
\name{create_directory_cnt}
\alias{create_directory_cnt}
\alias{create_directory_cnt.connector_fs}
\title{Create a directory}
\usage{
create_directory_cnt(connector_object, name, ...)

\method{create_directory_cnt}{connector_fs}(connector_object, name, ...)
}
\arguments{
\item{connector_object}{\link{connector} The connector object to use.}

\item{name}{\link{character} The name of the directory to create}

\item{...}{Additional arguments passed to the method for the individual connector.}
}
\value{
\link{invisible} connector_object.
}
\description{
Generic implementing of how to create a directory for a connector.
Mostly relevant for file storage connectors.

\itemize{
\item \link{connector_fs}: Uses \code{\link[=dir.create]{dir.create()}} to create a directory at the path of the connector.
}
}
\examples{
# Create a directory in a file storage

cnt <- connector_fs$new(tempdir())

cnt |>
  list_content_cnt(pattern = "new_folder")

cnt |>
  create_directory_cnt("new_folder")

# This will return new connector object of a newly created folder
new_connector <- cnt |>
  list_content_cnt(pattern = "new_folder")

cnt |>
  remove_directory_cnt("new_folder")

}
