% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_config.R
\name{datasources}
\alias{datasources}
\title{Previously used to extract data sources from connectors}
\usage{
datasources(connectors)
}
\arguments{
\item{connectors}{An object containing connectors with a "datasources" attribute.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}. Look for \verb{[list_datasources()]} instead.
}
