% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataTruncation.R
\name{DataTruncation}
\alias{DataTruncation}
\title{DataTruncation}
\usage{
DataTruncation(data, feature, truncTime = NULL, labels = NULL,
  save = FALSE, path = NULL)
}
\arguments{
\item{data}{CONNECTORList. (see \code{\link{DataImport}})}

\item{feature}{The column name reported in the AnnotationFile containing the feature  to be investigated.}

\item{truncTime}{An integer number corresponding to the time where  the curves will be truncated.}

\item{labels}{Vector containing the text for the title of axis names and plot title.}

\item{save}{If TRUE then the growth curves plot truncated at the "TruncTime" is saved into a pdf file.}

\item{path}{The folder path where the plot(s) will be saved. If it is missing, the plot is saved in the current working  directory.}
}
\value{
DataTruncation returns the line plot of growth curves with a vertical line at the truncation time and the CONNECTORList updated with the following elements: (i) a data frame with three variables (ID curves, volume and time values truncated at the chosen time), (ii) a vector collecting the number of truncated observations collected per sample, a data frame matching curses  with the chosen feature, (iv) the vector storing all the truncated time points of the samples (i.e. truncated time grid).
Furthermore, it prints an updated summary of the input data, updating the minimum and the maximum curve length after the data truncation.
}
\description{
Truncates the growth data at a specific time point chosen by the user.
}
\examples{

GrowDataFile<-"data/475dataset.xls"
AnnotationFile <-"data/475info.txt"

CONNECTORList <- DataImport(GrowDataFile,AnnotationFile)

CONNECTORList<- DataTruncation(CONNECTORList,"Progeny",truncTime=50,labels = c("time","volume","Tumor Growth"))

}
\author{
Cordero Francesca, Pernice Simone, Sirovich Roberta
}
