% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataVisualization.R
\name{DataVisualization}
\alias{DataVisualization}
\title{Data Visualitation}
\usage{
DataVisualization(data, feature, labels = NULL, save = FALSE,
  path = NULL)
}
\arguments{
\item{data}{CONNECTORList. (see \code{\link{DataImport}})}

\item{feature}{The column name reported in the AnnotationFile  containing the feature to be investigated.}

\item{labels}{The vector containing the text for the axis names and plot title.}

\item{save}{If TRUE then the plot is saved into a pdf file.}

\item{path}{The folder path  where the plot will be saved. If it is missing, the plot is saved in the current working  directory.}
}
\value{
Data Visualization returns a plot with the density time grid and the line plot of growth data as a ggplot object.
}
\description{
Computes the time grid and the line plot of growth data.
}
\examples{

GrowDataFile<-"data/1864dataset.xls"
AnnotationFile <-"data/1864info.txt"

CONNECTORList <- DataImport(GrowDataFile,AnnotationFile)

DataVisualization(CONNECTORList,"Progeny",labels = c("time","volume","Tumor Growth"))

}
\seealso{
\code{\link{GrowthCurve}}, code{\link{TimeGridDensity}}.
}
