% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterCoefficents.R
\name{DBindexL2dist}
\alias{DBindexL2dist}
\alias{L2dist.curve2mu}
\alias{L2dist.mu2mu}
\alias{L2dist.mu20}
\alias{Sclust.coeff}
\alias{Rclust.coeff}
\title{Functional Davies–Bouldin (fDB) index and L2 (or with derivatives) distances}
\usage{
L2dist.curve2mu(clust, fcm.curve, gauss.info, fcm.fit = NULL,
  deriv = 0)

L2dist.mu2mu(fcm.curve, gauss.info, fcm.fit = NULL, deriv = 0)

L2dist.mu20(fcm.curve, gauss.info, fcm.fit = NULL, deriv = 0)

Sclust.coeff(clust, fcm.curve, gauss.info, fcm.fit = NULL, deriv = 0)

Rclust.coeff(clust, fcm.curve, gauss.info, fcm.fit = NULL, deriv = 0)
}
\arguments{
\item{clust}{The vector reporting the cluster membership for each sample.}

\item{fcm.curve}{The list obtained applying the function fclust.curvepred to the fitfclust output, the FCM algorithm, more details in [Sugar and James].}

\item{gauss.info}{The list storing the (i) Gauss values, (ii) their corresponding time points, (iii) a and b, i.e. the lower and upper values for the integration.    (see \code{\link{gauss.quad}})}

\item{fcm.fit}{The fitfclust output, the FCM algorithm, more details in [Sugar and James].}

\item{deriv}{The derivative values (0, 1 or 2).}
}
\value{
The following indexes are calculated for obtaining a cluster separation measure:
\itemize{
\item{T:}{ the total tightness representing the dispersion measure given by
\deqn{ T = \sum_{k=1}^G \sum_{i=1}^n D_0(\hat{g}_i, \bar{g}^k); } (see \code{\link{L2dist.curve2mu}}) }
 \item{S_k:}{ 
 \deqn{ S_k = \sqrt{\frac{1}{G_k} \sum_{i=1}^{G_k} D_q^2(\hat{g}_i, \bar{g}^k);} }
 with G_k the number of curves in the k-th cluster; (see \code{\link{Sclust.coeff}})
 }
 \item{M_{hk}:}{ the distance between centroids (mean-curves) of h-th and k-th cluster 
 \deqn{M_{hk} =  D_q(\bar{g}^h, \bar{g}^k);  } (see\code{\link{L2dist.mu2mu}})
 }
 \item{R_{hk}:}{  a measure of how good the clustering is,
 \deqn{R_{hk} = \frac{S_h + S_k}{M_{hk}}; } (see\code{\link{Rclust.coeff}} )}
 \item{fDB_q:}{ functional Davies-Bouldin index, the cluster separation measure
 \deqn{fDB_q = \frac{1}{G} \sum_{k=1}^G \max_{h \neq k} {  \frac{S_h + S_k}{M_{hk}} }. }(see\code{\link{Rclust.coeff}}) }
}
Where the proximities measures choosen is defined as follow
 \deqn{D_q(f,g) = \sqrt( \integral | f^{(q)}(s)-g^{(q)}(s) |^2 ds ), d=0,1,2}
whit f and g are two curves and f^{(q)} and g^{(q)} are their q-th derivatives. Note that for q=0, the equation becomes the distance induced by the classical L^2-norm.

 @references
Gareth M. James and Catherine A. Sugar, (2003). Clustering for Sparsely Sampled Functional Data. Journal of the American Statistical Association.
}
\description{
L2dist.curve2mu calculates the L2 (or with derivatives) distance between all the spline estimated for each sample with respect to the corresponding cluster center curve. L2dist.mu2mu calculates the L2 distance between the cluster centers curves. L2dist.mu20 calculates the L2 distance between the cluster centers curves and zero. Sclust.coeff calculates the S coefficents for each cluster. Rclust.coeff calculates the R coefficents for each cluster.
}
\author{
Cordero Francesca, Pernice Simone, Sirovich Roberta
}
