% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscriminantPlot.R
\name{DiscriminantPlot}
\alias{DiscriminantPlot}
\title{Discriminant Plot}
\usage{
DiscriminantPlot(clusterdata, h, data, feature, save = FALSE, path = NULL)
}
\arguments{
\item{clusterdata}{Object belonging to the class funcyOutList if the model in study is the Functional Clustering Model (see \code{\link[funcy]{funcyOutList-class}}).}

\item{h}{The  number between 1 or 2 representing the dimension of the cluster mean space(see \code{\link{PCA.Analysis}}).}

\item{data}{CONNECTORList. (see \code{\link{DataImport}})}

\item{save}{If TRUE then the growth curves plot truncated at the ``truncTime'' is saved into a pdf file.}

\item{path}{The folder path where the plot(s) will be saved. If it is missing, the plot is saved in the current working  directory.

 @return}
}
\description{

}
\examples{


}
\references{
Gareth M. James and Catherine A. Sugar, (2003). Clustering for Sparsely Sampled Functional Data. Journal of the American Statistical Association.
}
\seealso{
\code{\link[funcy]{funcit}}.
}
