% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_generics.R, R/fs_methods.R
\name{download_directory_cnt}
\alias{download_directory_cnt}
\alias{download_directory_cnt.ConnectorFS}
\title{Download a directory}
\usage{
download_directory_cnt(connector_object, name, dir = name, ...)

\method{download_directory_cnt}{ConnectorFS}(connector_object, name, dir = basename(name), ...)
}
\arguments{
\item{connector_object}{\link{Connector} The connector object to use.}

\item{name}{\link{character} The name of the directory to download}

\item{dir}{\link{character} Path to the directory to download to}

\item{...}{Additional arguments passed to the method for the individual connector.}
}
\value{
\link{invisible} connector_object.
}
\description{
Generic implementing of how to download a directory for a connector.
Mostly relevant for file storage connectors.

\itemize{
\item \link{ConnectorFS}: Uses \code{\link[fs:copy]{fs::dir_copy()}}.
}
}
