% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtrapolationFunctions.R
\name{ExtrapolationFuncs}
\alias{ExtrapolationFuncs}
\alias{IndexesPlot.Extrapolation}
\alias{ConsMatrix.Extrapolation}
\alias{MostProbableClustering.Extrapolation}
\title{Extrapolate objects from ClusterAnalysis}
\usage{
IndexesPlot.Extrapolation(stability.list)

ConsMatrix.Extrapolation(stability.list)

MostProbableClustering.Extrapolation(stability.list, G)
}
\arguments{
\item{stability.list}{The list obtained from the ClusterAnalysis function. (see \code{\link{ClusterAnalysis}})}

\item{G}{The number of clusters.}
}
\description{
Extrapolate objects from ClusterAnalysis output list (see \code{\link{ClusterAnalysis}}).
}
\details{
\itemize{
\item{IndexesPlot.Extrapolation}{extrapolates from ClusterAnalysis output list the box plot fixing the h value.}
\item{ConsMatrix.Extrapolation}{extrapolates from ClusterAnalysis output list the Stability Matrix for G fixed.}
\item{MostProbableClustering.Extrapolation}{extrapolates from ClusterAnalysis output list the most frequent clustering among the several runs obtained  for G fixed.}
}
}
\author{
Cordero Francesca, Pernice Simone, Sirovich Roberta
}
