% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtrapolationFunctions.R, R/PlottingFunctions.R
\name{ExtrapolationFuncs}
\alias{ExtrapolationFuncs}
\alias{BoxPlot.Extrapolation}
\alias{ConsMatrix.Extrapolation}
\alias{MostProbableClustering.Extrapolation}
\alias{BoxPlot.ExtrapolationNew}
\alias{ConsMatrix.ExtrapolationNew}
\alias{MostProbableClustering.ExtrapolationNew}
\title{Extrapolate objects from StabilityAnalysis}
\usage{
BoxPlot.Extrapolation(stability.list, h)

ConsMatrix.Extrapolation(stability.list, h, G)

MostProbableClustering.Extrapolation(stability.list, h, G)

BoxPlot.ExtrapolationNew(stability.list)

ConsMatrix.ExtrapolationNew(stability.list, data)

MostProbableClustering.ExtrapolationNew(stability.list, G)
}
\arguments{
\item{stability.list}{The list obtained from the StabilityAnalysis function. (see \code{\link{StabilityAnalysis}})}

\item{h}{The number representing the dimension of the cluster mean space (see \code{\link{PCA.Analysis}}).}

\item{G}{The number of clusters.}
}
\description{
Extrapolate objects from StabilityAnalysis output list (see \code{\link{StabilityAnalysis}}).

Extrapolate objects from StabilityAnalysis output list (see \code{\link{StabilityAnalysis}}).
}
\details{
\itemize{
\item{BoxPlot.Extrapolation}{ extrapolates from StabilityAnalysis output list the box plot fixing the h value.}
\item{ConsMatrix.Extrapolation}{extrapolates from StabilityAnalysis output list the Consensus Matrix for G and h fixed.}
\item{MostProbableClustering.Extrapolation}{extrapolates from StabilityAnalysis output list the most probable cluster among the several runs obtained  for G and h fixed.}
}

\itemize{
\item{BoxPlot.Extrapolation}{ extrapolates from StabilityAnalysis output list the box plot fixing the h value.}
\item{ConsMatrix.Extrapolation}{extrapolates from StabilityAnalysis output list the Consensus Matrix for G and h fixed.}
\item{MostProbableClustering.Extrapolation}{extrapolates from StabilityAnalysis output list the most probable cluster among the several runs obtained  for G and h fixed.}
}
}
\author{
Cordero Francesca, Pernice Simone, Sirovich Roberta

Cordero Francesca, Pernice Simone, Sirovich Roberta
}
