% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{ExtrapolationNewFuncs}
\alias{ExtrapolationNewFuncs}
\alias{IndexesPlot.ExtrapolationNew}
\alias{ConsMatrix.ExtrapolationNew}
\alias{MostProbableClustering.ExtrapolationNew}
\title{Extrapolate objects from ClusterAnalysis}
\usage{
IndexesPlot.ExtrapolationNew(stability.list)

ConsMatrix.ExtrapolationNew(stability.list, data)

MostProbableClustering.ExtrapolationNew(stability.list, G)
}
\arguments{
\item{stability.list}{The list obtained from the ClusterAnalysis function. (see \code{\link{ClusterAnalysis}})}

\item{data}{CONNECTORList. (see \code{\link{DataImport}} or \code{\link{DataTruncation}})}

\item{G}{The number of clusters.}
}
\description{
Extrapolate objects from ClusterAnalysis output list (see \code{\link{ClusterAnalysis}}).
}
\details{
\itemize{
\item{IndexesPlot.Extrapolation}{extrapolates from ClusterAnalysis output list the box plot fixing the h value.}
\item{ConsMatrix.Extrapolation}{extrapolates from ClusterAnalysis output list the Consensus Matrix for G fixed.}
\item{MostProbableClustering.Extrapolation}{extrapolates from ClusterAnalysis output list the most frequent clustering among the several runs obtained  for G fixed.}
}
}
\author{
Cordero Francesca, Pernice Simone, Sirovich Roberta
}
