% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FittingAndClustering.R
\name{FittingAndClustering}
\alias{FittingAndClustering}
\title{Fitting and Clustering}
\usage{
FittingAndClustering(data, k, model, feature, fitting.method = "optimr",
  init = NULL, lower = NULL, upper = NULL, seed = NULL)
}
\arguments{
\item{data}{CONNECTORList.}

\item{k}{Number of clusters.}

\item{model}{Model name, i.e. Malthus, Gompertz and Logistic.}

\item{fitting.method}{}
}
\value{
A list containing per each curve (i) the parameters of the fitting, (ii) the centers, (iii) the cluster of affinity and (iv) the mean values.
}
\description{
Fitting and Clustering the cancer growth data with respect to the following growth functions: Malthus, Gompertz and Logistic. The cancer growth fitted functions are then clustered.
}
\examples{

}
