% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrowthCurve.R
\name{GrowthCurve}
\alias{GrowthCurve}
\title{Growth curves}
\usage{
GrowthCurve(data, feature, labels = NULL, save = FALSE, path = NULL)
}
\arguments{
\item{data}{CONNECTORList. (see \code{\link{DataImport}})}

\item{feature}{The column name reported in the AnnotationFile containing the feature to be investigated.}

\item{labels}{The vector containing the text for the axis names and plot title.}

\item{save}{If TRUE then the plot is saved into a pdf file.}

\item{path}{The folder path  where the plot will be saved. If it is missing, the plot is saved in the current working  directory.}
}
\value{
GrowthCurve returns a list containing the line plot as ggplot object, namely GrowthCurve_plot, and data, i.e. the CONNECTORList.
}
\description{
Generates the line plot of the growth data. The curves are colored with respect to the feature chosen by the user and/or reported in the AnnotationFile.
}
\examples{
GrowDataFile<-"data/745dataset.xls"
AnnotationFile <-"data/745info.txt"

CONNECTORList <- DataImport(GrowDataFile,AnnotationFile)

GrowthCurve(CONNECTORLis,"Progeny",labels=c("Time","Volume","Tumor Growth"))

}
