% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrowthCurve.R
\name{GrowthCurve}
\alias{GrowthCurve}
\title{Growth curves}
\usage{
GrowthCurve(data, feature, labels = NULL)
}
\arguments{
\item{data}{CONNECTORList. (see \code{\link{DataImport}})}

\item{feature}{The column name reported in the AnnotationFile containing the feature to be investigated.}

\item{labels}{The vector containing the text for the axis names and plot title.}
}
\value{
GrowthCurve returns a list containing the line plot and the CONNECTORList.
}
\description{
Generates the line plot of the growth data. The curves are colored with respect to the feature chosen by the user and/or reported in the AnnotationFile.
}
\examples{
GrowDataFile<-"data/1864dataset.xls"
AnnotationFile <-"data/1864info.txt"

CONNECTORList <- DataImport(GrowDataFile,AnnotationFile)

GrowthCurve(CONNECTORLis,"Progeny",labels=c("time","volume","Tumor Growth"))

}
