% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haus.R
\name{hausdorff}
\alias{hausdorff}
\title{Hausdorff distance}
\usage{
hausdorff(A, B)
}
\arguments{
\item{A}{A set.}

\item{B}{A set.}
}
\value{
Returnst the Hausdorff distance between A and B.
}
\description{
Calculates the Hausdorff distance between two sets.
}
