% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaximumDiscriminationFunction.R
\name{MaximumDiscriminationFunction}
\alias{MaximumDiscriminationFunction}
\title{Maximum Discrimination Function}
\usage{
MaximumDiscriminationFunction(clusterdata, absvalue = TRUE)
}
\arguments{
\item{clusterdata}{Object belonging to the class funcyOutList if the model in study is the Functional Clustering Model (see \code{\link[funcy]{funcyOutList-class}}).}

\item{absvalue}{@return}
}
\description{

}
\examples{


}
\references{
Gareth M. James and Catherine A. Sugar, (2003). Clustering for Sparsely Sampled Functional Data. Journal of the American Statistical Association.
}
