% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_connectors.R
\name{nested_connectors}
\alias{nested_connectors}
\title{Create a nested connectors object}
\usage{
nested_connectors(...)
}
\arguments{
\item{...}{Any number of named \code{\link[=connectors]{connectors()}} objects.}
}
\value{
A list with class "nested_connectors" containing the provided arguments.
}
\description{
This function creates a nested connectors object from the provided arguments.
}
\examples{
nested_connectors(
  trial_1 = connectors(
    sdtm = connector_fs(path = tempdir())
  ),
  trial_2 = connectors(
    sdtm = connector_dbi(drv = RSQLite::SQLite())
  )
)

}
