% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.Analysis.R
\name{PCA.Analysis}
\alias{PCA.Analysis}
\title{PCA Analysis}
\usage{
PCA.Analysis(data, p = 5, save = FALSE, path = NULL)
}
\arguments{
\item{data}{CONNECTORList.  (see \code{\link{DataImport}})}

\item{p}{The dimension of the natural cubic spline basis.}

\item{save}{If TRUE the bar plot of the PCA components is saved into a pdf file.}

\item{path}{The folder path   where the plot will be saved. If it is missing, the plot is saved in the current working  directory.}
}
\value{
PCA Analysis returns the bar plot of the PCA components and the vector of percentage values associated with the PCA components.
}
\description{
Generates a bar plot reporting the percentage values associated with each component identified by the Principal Component Analysis (PCA). Considering the Functional Clustering Model [Sugar and James] the PCA is applied on the natural cubic spline coefficients to estimate the dimension of the subspace, h, in which the cluster means lie. This is due to a parameterization of the cluster means for producing low-dimensional representations of the curves, for this reason \eqn{h\le p}.
}
\examples{

GrowDataFile<-"data/1864dataset.xls"
AnnotationFile <-"data/1864info.txt"

CONNECTORList <- DataImport(GrowDataFile,AnnotationFile)

CONNECTORList<- DataTruncation(CONNECTORList,"Progeny",truncTime=60,labels = c("time","volume","Tumor Growth"))

PCA<-PCA.Analysis(CONNECTORList)
PCA$plot

}
\references{
Gareth M. James and Catherine A. Sugar, (2003). Clustering for Sparsely Sampled Functional Data. Journal of the American Statistical Association.
}
\seealso{
\code{\link[fda]{create.bspline.irregular}}.
}
