% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSeparationTightness.R
\name{PlotSeparationTightness}
\alias{PlotSeparationTightness}
\title{Separation and Tightness Visualization}
\usage{
PlotSeparationTightness(clusterdata, Title = NULL, save = FALSE,
  path = NULL)
}
\arguments{
\item{clusterdata}{Object belonging to the class funcyOutList if the model in study is the Functional Clustering Model (see \code{\link[funcy]{funcyOutList-class}}). Otherwise a list derived from fitting and clustering the data using Malthus, Gompertz or Logistic model storing the parameters and the cluster membership for each sample, the  parameters of the center and the mean curve values for each cluster (see \code{\link{FittingAndClustering}}).}

\item{Title}{Title of the Separation and Tightness plot.}

\item{save}{If TRUE then the Separation and Tightness plot is saved into a pdf file.}

\item{path}{The folder path where the plots will be saved. If it is missing, the plots are saved in the current working  directory.}
}
\value{
Returns the Separation and Tightness plot.
}
\description{
Visualizes the plots reporting the Separation and Tightness clustering measures.
}
\details{
To visualize the Separation (how much the cluster mean curves are different from each other)  and the Tightness (how much the curves belonging to the same cluster are different from the corresponding mean curve) of the clusters three circles are drawn according to the mean and variance of the Hausdorff distance of fitted curves within a cluster with respect to the cluster mean curve. In each circle the number represent the cluster mean curves while the symbols represent the objects in the cluster, one symbol for each cluster.
}
\examples{
### Data files
GrowDataFile<-"data/1864dataset.xls"
AnnotationFile <-"data/1864info.txt"

### Merge curves and target file
CONNECTORList <- DataImport(GrowDataFile,AnnotationFile)

### Truncation

CONNECTORList<- DataTruncation(CONNECTORList,feature="Progeny",60,labels = c("time","volume","Tumor Growth"))

############################## FCM ##############################

CONNECTORList.FCM <- ClusterChoice(CONNECTORList,k=c(2:6),h=2)
CONNECTORList.FCM.k4.h2<- CONNECTORList.FCM$FCM_all$`k= 4`$`h= 2`

PlotSeparationTightness(CONNECTORList.FCM.k4.h2,Title = "FCM Cluster betweenness and withinness")

############################## MALTHUS ##############################
lower<-c(10^(-5),0)
upper<-c(10^2,10^3)
init<- list(V0=max(0.1,min(CONNECTORList$Dataset$Vol)),a=1)


Malthus1<- FittingAndClustering(data = CONNECTORList, k = 4, model="Malthus",feature="Progeny",fitting.method="optimr",lower=lower,upper=upper,init=init)


PlotSeparationTightness(Malthus1,Title = "Malthus Cluster betweenness and withinness")

}
