% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_read.R
\name{read_ext}
\alias{read_ext}
\alias{read_ext.default}
\alias{read_ext.txt}
\alias{read_ext.csv}
\alias{read_ext.parquet}
\alias{read_ext.rds}
\alias{read_ext.sas7bdat}
\alias{read_ext.xpt}
\alias{read_ext.yml}
\title{Read a file based on this extension}
\usage{
read_ext(path, ...)

\method{read_ext}{default}(path, ...)

\method{read_ext}{txt}(path, ...)

\method{read_ext}{csv}(path, ...)

\method{read_ext}{parquet}(path, ...)

\method{read_ext}{rds}(path, ...)

\method{read_ext}{sas7bdat}(path, ...)

\method{read_ext}{xpt}(path, ...)

\method{read_ext}{yml}(path, eval.expr = TRUE, ...)
}
\arguments{
\item{path}{Path to the file}

\item{...}{Other parameters for read's functions}
}
\value{
the result of the reading function
}
\description{
Read a file based on this extension

Default method for reading files

For txt files

For CSV files

For parquet files

For RDS files

For SAS files

For XPT files

For yaml files
}
\examples{
temp_csv <- tempfile("iris", fileext = ".csv")
write.csv(iris, temp_csv, row.names = FALSE)
class(temp_csv) <- "csv"
read_ext(temp_csv)
temp_txt <- tempfile("iris", fileext = ".txt")
write.csv(iris, temp_txt, row.names = FALSE)
class(temp_txt) <- "txt"
read_ext(temp_txt)
}
