% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{read_file}
\alias{read_file}
\title{Read files based on the extension}
\usage{
read_file(path, ...)
}
\arguments{
\item{path}{Path to the file}

\item{...}{Other parameters for read's functions}
}
\value{
the result of the reading function
}
\description{
The aim of this function is to identify the extension on the file to dispatch it.
}
\examples{
temp_csv <- tempfile("iris", fileext = ".csv")
write.csv(iris, temp_csv, row.names = FALSE)
read_file(temp_csv)
}
