% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_connector.R
\name{use_connector}
\alias{use_connector}
\title{Use connector}
\usage{
use_connector(config_file = "_connector.yml")
}
\arguments{
\item{config_file}{Path to the connector config file, relative to the project}
}
\description{
Utility function to setup connections with connector in your project:

Creates configuration file (default \verb{_connector.yml})

See \code{vignette("connector")} for how to configure the file.
}
