% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliarSeparationAndTightness.R
\name{WithCluster_MeanDist}
\alias{WithCluster_MeanDist}
\title{Computes the Hausdorff distance between the curves belonging to the i-th cluster and the corresponding mean curve.}
\usage{
WithCluster_MeanDist(ClustCurve, MeanCurves, i)
}
\description{
Computes the Hausdorff distance between the curves belonging to the i-th cluster and the corresponding mean curve.
}
