% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conts_datasources.R
\name{write_datasources}
\alias{write_datasources}
\title{Write datasources attribute into a config file}
\usage{
write_datasources(connectors, file)
}
\arguments{
\item{connectors}{An object containing connectors with a "datasources" attribute.}

\item{file}{path to the config file}
}
\value{
A config file with the datasources attribute and can be reuse in the connect function
}
\description{
Reproduce your workflow by creating a config file based on a connectors object and his datacrouces attribut
}
\examples{

# Connect to the datasources specified in it
config <- system.file("config", "default_config.yml", package = "connector")
cnts <- connect(config)

# Extract the datasources to a config file
yml_file <- tempfile(fileext = ".yml")
write_datasources(cnts, yml_file)

# Reconnect using the new config file
re_connect <- connect(yml_file)
re_connect

}
