% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_write.R
\name{write_ext}
\alias{write_ext}
\alias{write_ext.csv}
\alias{write_ext.parquet}
\alias{write_ext.rds}
\alias{write_ext.xpt}
\title{Write a file based on this extension}
\usage{
write_ext(file, x, ...)

\method{write_ext}{csv}(file, x, ...)

\method{write_ext}{csv}(file, x, ...)

\method{write_ext}{parquet}(file, x, ...)

\method{write_ext}{rds}(file, x, ...)

\method{write_ext}{xpt}(file, x, ...)
}
\arguments{
\item{file}{Path to write the file}

\item{x}{Object to write}

\item{...}{Other parameters for write's functions}
}
\value{
the result of the Writing function
}
\description{
Write a file based on this extension

For txt files

For CSV files

For parquet files

For RDS files

For xpt files
}
\examples{
temp_csv <- tempfile("iris", fileext = ".csv")
class(temp_csv) <- "csv"
write_ext(temp_csv, iris)
}
