% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{write_file}
\alias{write_file}
\title{Write a file based on this extension}
\usage{
write_file(x, file, ...)
}
\arguments{
\item{x}{Object to write}

\item{file}{Path to write the file}

\item{...}{Other parameters for write's functions}
}
\value{
the result of the Writing function
}
\description{
Write a file based on this extension
}
\examples{
temp_csv <- tempfile("iris", fileext = ".csv")
write_file(iris, temp_csv)
}
